refactor(components): decompose large page/layout files into focused sub-components (#82)
- Split PropertyDetailPage (2795→606 lines) into 6 tab files under pages/PropertyDetail/
- Split Layout (733→486 lines) into ActivityFeedDrawer, UserMenuPopover, and activityFeed service
- Split DashboardPage (1856→1717 lines) into ScoreSparkline, ScoreHistoryChart, PropertyCard
- Split LandingPage (1643→765 lines) by extracting CSS block to landingStyles.ts
- Fix referralService TODO format to TODO(#82)

All 2910 unit tests pass, TypeScript clean.
