+++ frontend/src/pages/PropertyDetailPage.tsx
@@ -1,13 +1,12 @@
 import React, { useEffect, useState } from "react";
 import { useParams, useNavigate, useSearchParams } from "react-router-dom";
-import { ArrowLeft, Share2, Shield, Wrench, MessageSquare, Calendar, DollarSign, AlertCircle, Star, Upload, Zap, Trash2, TrendingUp, ExternalLink, PhoneCall } from "lucide-react";
+import { ArrowLeft, Share2, Shield, Wrench, MessageSquare, AlertCircle } from "lucide-react";
 import { Layout } from "@/components/Layout";
 import { Button } from "@/components/Button";
 import { Badge } from "@/components/Badge";
 import { GenerateReportModal }     from "@/components/GenerateReportModal";
 import { LogJobModal }              from "@/components/LogJobModal";
 import { RequestQuoteModal }        from "@/components/RequestQuoteModal";
-import { AddRoomModal }             from "@/components/AddRoomModal";
 import { InviteContractorModal }    from "@/components/InviteContractorModal";
 import { propertyService, Property } from "@/services/property";
 import { jobService, Job } from "@/services/job";
@@ -22,11 +21,8 @@ import { getRecentScoreEvents, type ScoreEvent } from "@/services/scoreEventServ
 import { getReEngagementPrompts, type ReEngagementPrompt } from "@/services/reEngagementService";
 import { marketService, jobToSummary, type PropertyProfile, type ProjectRecommendation } from "@/services/market";
 import { getWeeklyPulse } from "@/services/pulseService";
-import { roomService, Room as RoomRecord, type UpdateRoomArgs, type AddFixtureArgs, type Fixture } from "@/services/room";
+import { roomService, type Room as RoomRecord } from "@/services/room";
 import { paymentService, type PlanTier } from "@/services/payment";
-import { billService, extractBill, TierLimitReachedError, type BillRecord, type BillExtraction, type BillType } from "@/services/billService";
-import { getUsageTrend, analyzeEfficiencyTrend, findRebates, negotiateTelecom, type RebateResult, type TelecomNegotiationResult } from "@/services/billsIntelligence";
-import { UpgradeGate } from "@/components/UpgradeGate";
 import { ScorePanel } from "@/components/ScorePanel";
 import { ScoreActivityFeed } from "@/components/ScoreActivityFeed";
 import { AlertStack } from "@/components/AlertStack";
@@ -38,6 +34,12 @@ import FsboPanel from "@/components/FsboPanel";
 import { usePropertyStore } from "@/store/propertyStore";
 import { useAuthStore } from "@/store/authStore";
 import toast from "react-hot-toast";
+import { TimelineTab }  from "./PropertyDetail/TimelineTab";
+import { JobsTab }      from "./PropertyDetail/JobsTab";
+import { DocumentsTab } from "./PropertyDetail/DocumentsTab";
+import { SettingsTab }  from "./PropertyDetail/SettingsTab";
+import { RoomsTab }     from "./PropertyDetail/RoomsTab";
+import { BillsTab }     from "./PropertyDetail/BillsTab";
 
 import { COLORS, FONTS, RADIUS, SHADOWS } from "@/theme";
 
@@ -602,2194 +604,3 @@ export default function PropertyDetailPage() {
   );
 }
 
-function SigPill({ signed, label }: { signed: boolean; label: string }) {
-  return (
-    <span style={{
-      display: "inline-flex", alignItems: "center", gap: "0.25rem",
-      fontFamily: FONTS.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase",
-      padding: "0.15rem 0.625rem", borderRadius: 100,
-      border: `1px solid ${signed ? COLORS.sageMid : COLORS.rule}`,
-      color: signed ? COLORS.sage : COLORS.plumMid,
-      background: signed ? COLORS.sageLight : COLORS.white,
-    }}>
-      {signed ? "✓" : "○"} {label}
-    </span>
-  );
-}
-
-function PhotoStrip({ photos, jobId, onUpload }: { photos: Photo[]; jobId: string; onUpload: (jobId: string, file: File) => void }) {
-  const LS = { rule: COLORS.rule, inkLight: COLORS.plumMid, rust: COLORS.sage, mono: FONTS.mono };
-  const inputRef = React.useRef<HTMLInputElement>(null);
-  const [lightboxIdx, setLightboxIdx] = useState<number | null>(null);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) { onUpload(jobId, file); e.target.value = ""; }
-  };
-
-  const openLightbox = (idx: number) => setLightboxIdx(idx);
-  const closeLightbox = () => setLightboxIdx(null);
-  const prev = (e: React.MouseEvent) => { e.stopPropagation(); setLightboxIdx((i) => i !== null ? Math.max(0, i - 1) : null); };
-  const next = (e: React.MouseEvent) => { e.stopPropagation(); setLightboxIdx((i) => i !== null ? Math.min(photos.length - 1, i + 1) : null); };
-
-  // Keyboard navigation
-  React.useEffect(() => {
-    if (lightboxIdx === null) return;
-    const handler = (e: KeyboardEvent) => {
-      if (e.key === "ArrowLeft")  setLightboxIdx((i) => i !== null ? Math.max(0, i - 1) : null);
-      if (e.key === "ArrowRight") setLightboxIdx((i) => i !== null ? Math.min(photos.length - 1, i + 1) : null);
-      if (e.key === "Escape")     setLightboxIdx(null);
-    };
-    window.addEventListener("keydown", handler);
-    return () => window.removeEventListener("keydown", handler);
-  }, [lightboxIdx, photos.length]);
-
-  const activePh = lightboxIdx !== null ? photos[lightboxIdx] : null;
-
-  return (
-    <>
-      <div style={{ display: "flex", alignItems: "center", gap: "0.5rem", marginTop: "0.75rem", flexWrap: "wrap" }}>
-        {photos.slice(0, 5).map((p, i) => (
-          <img
-            key={p.id}
-            src={p.url}
-            alt={p.description}
-            title={p.description}
-            onClick={() => openLightbox(i)}
-            style={{ width: 48, height: 48, objectFit: "cover", border: `1px solid ${LS.rule}`, cursor: "pointer" }}
-          />
-        ))}
-        {photos.length > 5 && (
-          <button
-            onClick={() => openLightbox(5)}
-            style={{ fontFamily: LS.mono, fontSize: "0.6rem", color: LS.inkLight, background: "none", border: `1px solid ${LS.rule}`, padding: "0.2rem 0.5rem", cursor: "pointer" }}
-          >
-            +{photos.length - 5} more
-          </button>
-        )}
-        <button
-          onClick={() => inputRef.current?.click()}
-          style={{
-            padding: "0.2rem 0.6rem",
-            fontFamily: LS.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase",
-            color: LS.inkLight, background: "none", border: `1px solid ${LS.rule}`, cursor: "pointer",
-          }}
-        >
-          + Add Photo
-        </button>
-        <input ref={inputRef} type="file" accept="image/*" style={{ display: "none" }} onChange={handleChange} />
-      </div>
-
-      {/* Lightbox overlay */}
-      {activePh && (
-        <div
-          onClick={closeLightbox}
-          style={{
-            position: "fixed", inset: 0, background: "rgba(14,14,12,0.92)",
-            display: "flex", alignItems: "center", justifyContent: "center",
-            zIndex: 9999, padding: "2rem",
-          }}
-        >
-          {/* Prev */}
-          <button
-            onClick={prev}
-            disabled={lightboxIdx === 0}
-            style={{ position: "absolute", left: "1.5rem", top: "50%", transform: "translateY(-50%)", background: "none", border: `1px solid rgba(255,255,255,0.3)`, color: COLORS.white, padding: "0.75rem", cursor: lightboxIdx === 0 ? "default" : "pointer", opacity: lightboxIdx === 0 ? 0.3 : 1, fontSize: "1.25rem", lineHeight: 1 }}
-          >
-            ‹
-          </button>
-
-          {/* Image */}
-          <div onClick={(e) => e.stopPropagation()} style={{ display: "flex", flexDirection: "column", alignItems: "center", maxWidth: "80vw", maxHeight: "80vh" }}>
-            <img
-              src={activePh.url}
-              alt={activePh.description}
-              style={{ maxWidth: "100%", maxHeight: "70vh", objectFit: "contain", border: "1px solid rgba(255,255,255,0.2)" }}
-            />
-            <div style={{ marginTop: "0.75rem", display: "flex", alignItems: "center", gap: "1rem" }}>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.08em", color: "rgba(255,255,255,0.6)" }}>
-                {activePh.description || "No description"}
-              </span>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.55rem", color: "rgba(255,255,255,0.4)" }}>
-                {(lightboxIdx ?? 0) + 1} / {photos.length}
-              </span>
-            </div>
-          </div>
-
-          {/* Next */}
-          <button
-            onClick={next}
-            disabled={lightboxIdx === photos.length - 1}
-            style={{ position: "absolute", right: "1.5rem", top: "50%", transform: "translateY(-50%)", background: "none", border: `1px solid rgba(255,255,255,0.3)`, color: COLORS.white, padding: "0.75rem", cursor: lightboxIdx === photos.length - 1 ? "default" : "pointer", opacity: lightboxIdx === photos.length - 1 ? 0.3 : 1, fontSize: "1.25rem", lineHeight: 1 }}
-          >
-            ›
-          </button>
-
-          {/* Close */}
-          <button
-            onClick={closeLightbox}
-            style={{ position: "absolute", top: "1.25rem", right: "1.25rem", background: "none", border: `1px solid rgba(255,255,255,0.3)`, color: COLORS.white, padding: "0.375rem 0.75rem", fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", cursor: "pointer" }}
-          >
-            Close
-          </button>
-        </div>
-      )}
-    </>
-  );
-}
-
-function warrantyStatus(job: Job): { label: string; color: string; bg: string } | null {
-  if (!job.warrantyMonths || job.warrantyMonths <= 0) return null;
-  const jobDate = new Date(job.date).getTime();
-  const expiryMs = jobDate + job.warrantyMonths * 30.44 * 24 * 60 * 60 * 1000;
-  const now = Date.now();
-  const daysLeft = Math.round((expiryMs - now) / (24 * 60 * 60 * 1000));
-  if (daysLeft < 0) return { label: "Warranty expired", color: COLORS.plumMid, bg: COLORS.white };
-  if (daysLeft <= 90) return { label: `Warranty: ${daysLeft}d left`, color: COLORS.sage, bg: COLORS.blush };
-  const monthsLeft = Math.round(daysLeft / 30);
-  return { label: `Warranty: ${monthsLeft}mo left`, color: COLORS.sage, bg: COLORS.sageLight };
-}
-
-function TimelineTab({ property, jobs, onVerify, currentPrincipal, photosByJob, onPhotoUpload, onInviteContractor }: {
-  property: Property;
-  jobs: Job[];
-  onVerify: (id: string) => void;
-  currentPrincipal: string | null;
-  photosByJob: Record<string, Photo[]>;
-  onPhotoUpload: (jobId: string, file: File) => void;
-  onInviteContractor: (job: Job) => void;
-}) {
-  const S = { ink: COLORS.plum, rule: COLORS.rule, rust: COLORS.sage, inkLight: COLORS.plumMid, sage: COLORS.sage, mono: FONTS.mono, serif: FONTS.serif };
-  const navigate = useNavigate();
-  const [justVerified,        setJustVerified]        = React.useState<string | null>(null);
-  const [reviewNudgeJob,      setReviewNudgeJob]      = React.useState<Job | null>(null);
-  const [newestFirst,         setNewestFirst]         = React.useState(true);
-  const [expandedJobId,       setExpandedJobId]       = React.useState<string | null>(null);
-  const [warrantyUploading,   setWarrantyUploading]   = React.useState<string | null>(null); // jobId being uploaded
-  const warrantyInputRefs = React.useRef<Record<string, HTMLInputElement | null>>({});
-
-  const handleWarrantyUpload = async (job: Job, file: File) => {
-    setWarrantyUploading(job.id);
-    try {
-      await photoService.upload(file, job.id, String(property.id), "Warranty", `Warranty|${job.serviceType}|${file.name}`);
-      toast.success("Warranty document uploaded");
-    } catch (err: any) {
-      const msg: string = err.message ?? "Upload failed";
-      toast.error(msg === "Duplicate" ? "Already uploaded" : msg);
-    } finally {
-      setWarrantyUploading(null);
-    }
-  };
-
-  const verifiedCount = jobs.filter((j) => j.verified).length;
-
-  // Sort jobs and group by year for the timeline
-  const sortedJobs = React.useMemo(() => {
-    return [...jobs].sort((a, b) =>
-      newestFirst ? b.date.localeCompare(a.date) : a.date.localeCompare(b.date)
-    );
-  }, [jobs, newestFirst]);
-
-  const handleVerify = (jobId: string) => {
-    onVerify(jobId);
-    setJustVerified(jobId);
-    const job = jobs.find((j) => j.id === jobId);
-    if (job && !job.isDiy) setReviewNudgeJob(job);
-    setTimeout(() => setJustVerified(null), 2500);
-  };
-
-  if (jobs.length === 0) {
-    return (
-      <div style={{ border: `1px dashed ${S.rule}`, padding: "3rem", textAlign: "center" }}>
-        <Calendar size={36} color={S.rule} style={{ margin: "0 auto 1rem" }} />
-        <p style={{ fontFamily: S.serif, fontWeight: 700, marginBottom: "0.375rem" }}>No jobs recorded yet</p>
-        <p style={{ fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.06em", color: S.inkLight }}>
-          Log your first maintenance job to start the timeline.
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <>
-      {/* 3-service engagement milestone */}
-      {verifiedCount >= 3 && (
-        <div style={{ border: `1px solid ${S.sage}`, background: COLORS.sageLight, padding: "0.875rem 1.25rem", marginBottom: "1rem", display: "flex", alignItems: "center", gap: "0.75rem" }}>
-          <span style={{ fontSize: "1.25rem" }}>🏅</span>
-          <div>
-            <p style={{ fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.1em", textTransform: "uppercase", color: S.sage, marginBottom: "0.1rem" }}>
-              Home History Taking Shape
-            </p>
-            <p style={{ fontSize: "0.8rem", color: S.inkLight, fontWeight: 300 }}>
-              {verifiedCount} verified jobs on-chain. Your HomeGentic report is ready to impress buyers.
-            </p>
-          </div>
-        </div>
-      )}
-
-      {/* Review nudge — shown after a contractor job is verified */}
-      {reviewNudgeJob && (
-        <div style={{
-          display: "flex", alignItems: "center", justifyContent: "space-between", gap: "1rem",
-          border: `1px solid ${COLORS.sageMid}`, padding: "0.875rem 1.25rem", marginBottom: "1rem",
-          background: COLORS.butter, flexWrap: "wrap",
-        }}>
-          <div style={{ display: "flex", alignItems: "center", gap: "0.625rem" }}>
-            <Star size={14} color={COLORS.plum} style={{ flexShrink: 0 }} />
-            <div>
-              <p style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", color: COLORS.plum, marginBottom: "0.15rem" }}>
-                Job verified — leave a review
-              </p>
-              <p style={{ fontSize: "0.8rem", fontWeight: 300, color: COLORS.plumMid }}>
-                Help other homeowners by reviewing {reviewNudgeJob.contractorName || "this contractor"}.
-              </p>
-            </div>
-          </div>
-          <div style={{ display: "flex", alignItems: "center", gap: "0.5rem" }}>
-            {reviewNudgeJob.contractor && (
-              <button
-                onClick={() => navigate(`/contractor/${reviewNudgeJob.contractor}`)}
-                style={{
-                  fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase",
-                  padding: "0.375rem 0.875rem", border: `1px solid ${COLORS.sageMid}`, color: COLORS.plum,
-                  background: "none", cursor: "pointer",
-                }}
-              >
-                Leave a Review
-              </button>
-            )}
-            <button
-              onClick={() => setReviewNudgeJob(null)}
-              style={{ background: "none", border: "none", cursor: "pointer", color: COLORS.plumMid, padding: "0.25rem" }}
-            >
-              ×
-            </button>
-          </div>
-        </div>
-      )}
-
-      {/* Timeline sort control */}
-      <div style={{ display: "flex", justifyContent: "flex-end", marginBottom: "0.75rem" }}>
-        <button
-          onClick={() => setNewestFirst((v) => !v)}
-          style={{
-            fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase",
-            padding: "0.25rem 0.75rem", border: `1px solid ${S.rule}`, background: "none",
-            color: S.inkLight, cursor: "pointer",
-          }}
-        >
-          {newestFirst ? "Newest First ↓" : "Oldest First ↑"}
-        </button>
-      </div>
-
-      {/* Visual timeline */}
-      <div style={{ paddingLeft: "1.5rem", position: "relative" }}>
-        {/* Vertical spine */}
-        <div style={{ position: "absolute", left: "0.5rem", top: 0, bottom: 0, width: "1px", background: S.rule }} />
-
-        {sortedJobs.map((job, idx) => {
-          const isHomeowner  = currentPrincipal && job.homeowner === currentPrincipal;
-          const canSign      = !job.verified && isHomeowner && !job.homeownerSigned;
-          const needsBothSig = !job.isDiy;
-          const isFlashing   = justVerified === job.id || (job.verified && justVerified === job.id);
-          const warranty     = warrantyStatus(job);
-          const year         = job.date.slice(0, 4);
-          const prevYear     = idx > 0 ? sortedJobs[idx - 1].date.slice(0, 4) : null;
-          const showYearMark = year !== prevYear;
-
-          return (
-            <React.Fragment key={job.id}>
-              {/* Year separator */}
-              {showYearMark && (
-                <div style={{ display: "flex", alignItems: "center", gap: "0.75rem", marginBottom: "0.5rem", marginTop: idx > 0 ? "1.25rem" : 0 }}>
-                  <div style={{
-                    position: "absolute", left: "0.125rem",
-                    width: "0.75rem", height: "0.75rem", background: S.ink,
-                    display: "flex", alignItems: "center", justifyContent: "center",
-                  }} />
-                  <span style={{ fontFamily: S.serif, fontWeight: 900, fontSize: "1rem", color: S.ink, marginLeft: "0.25rem" }}>
-                    {year}
-                  </span>
-                </div>
-              )}
-
-              {/* Timeline entry */}
-              <div style={{ position: "relative", marginBottom: "1px" }}>
-                {/* Dot on spine */}
-                <div style={{
-                  position: "absolute", left: "-1.25rem", top: "1.375rem",
-                  width: "0.5rem", height: "0.5rem",
-                  background: job.verified ? S.sage : S.rule,
-                  border: `1px solid ${job.verified ? S.sage : S.inkLight}`,
-                }} />
-
-                <div
-                  data-testid={`job-${job.serviceType.toLowerCase().replace(/\s+/g, "-")}`}
-                  style={{
-                    background: isFlashing ? COLORS.sageLight : COLORS.white,
-                    padding: "1.25rem",
-                    border: `1px solid ${S.rule}`,
-                    transition: "background 0.6s ease",
-                  }}
-                >
-                  <div style={{ display: "flex", justifyContent: "space-between", alignItems: "flex-start" }}>
-                    <div>
-                      <p style={{ fontWeight: 500, fontSize: "0.875rem", marginBottom: "0.125rem" }}>{job.serviceType}</p>
-                      <p style={{ fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.06em", color: S.inkLight }}>
-                        {job.isDiy ? "DIY" : job.contractorName} · {job.date}
-                      </p>
-                    </div>
-                    <div style={{ textAlign: "right" }}>
-                      <p style={{ fontFamily: S.mono, fontSize: "0.875rem", fontWeight: 500, marginBottom: "0.25rem" }}>
-                        ${(job.amount / 100).toLocaleString()}
-                      </p>
-                      <Badge variant={job.status === "verified" ? "success" : job.status === "completed" ? "info" : "warning"} size="sm">
-                        {isFlashing ? "⛓ locked on-chain" : job.status}
-                      </Badge>
-                    </div>
-                  </div>
-
-                  {job.description && (
-                    <p style={{ fontSize: "0.8rem", color: S.inkLight, fontWeight: 300, marginTop: "0.5rem" }}>{job.description}</p>
-                  )}
-
-                  {/* Warranty pill */}
-                  {warranty && (
-                    <div style={{ marginTop: "0.5rem" }}>
-                      <span style={{
-                        display: "inline-flex", alignItems: "center", gap: "0.25rem",
-                        fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase",
-                        padding: "0.15rem 0.5rem",
-                        color: warranty.color, background: warranty.bg,
-                        border: `1px solid ${warranty.color}40`,
-                      }}>
-                        🛡 {warranty.label}
-                      </span>
-                    </div>
-                  )}
-
-                  {/* Signature status */}
-                  {!job.verified && (
-                    <div style={{ display: "flex", alignItems: "center", gap: "0.5rem", marginTop: "0.75rem", flexWrap: "wrap" }}>
-                      <SigPill signed={job.homeownerSigned} label="Homeowner" />
-                      {needsBothSig && (
-                        <SigPill signed={job.contractorSigned} label={job.contractor ? "Contractor" : "Contractor (not linked)"} />
-                      )}
-                      {canSign && (
-                        <button
-                          onClick={() => handleVerify(job.id)}
-                          style={{
-                            padding: "0.25rem 0.75rem",
-                            fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase",
-                            color: S.rust, background: "none", border: `1px solid ${S.rust}`, cursor: "pointer",
-                          }}
-                        >
-                          Sign →
-                        </button>
-                      )}
-                      {job.homeownerSigned && !job.contractorSigned && !job.isDiy && !job.verified && (
-                        <>
-                          <span style={{ fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight }}>
-                            Awaiting contractor signature
-                          </span>
-                          <button
-                            onClick={() => onInviteContractor(job)}
-                            style={{
-                              padding: "0.25rem 0.75rem",
-                              fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase",
-                              color: S.sage, background: "none", border: `1px solid ${S.sage}`, cursor: "pointer",
-                            }}
-                          >
-                            Invite →
-                          </button>
-                        </>
-                      )}
-                    </div>
-                  )}
-
-                  {/* Detail expand toggle */}
-                  {(job.permitNumber || job.description) && (
-                    <button
-                      onClick={() => setExpandedJobId((prev) => prev === job.id ? null : job.id)}
-                      style={{ display: "inline-flex", alignItems: "center", gap: "0.25rem", marginTop: "0.625rem", fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, background: "none", border: "none", cursor: "pointer", padding: 0 }}
-                    >
-                      {expandedJobId === job.id ? "▲ less" : "▼ details"}
-                    </button>
-                  )}
-
-                  {expandedJobId === job.id && (
-                    <div style={{ marginTop: "0.625rem", padding: "0.75rem", background: COLORS.white, border: `1px solid ${S.rule}`, display: "flex", flexDirection: "column", gap: "0.5rem" }}>
-                      {job.permitNumber && (
-                        <div style={{ display: "flex", gap: "0.75rem" }}>
-                          <span style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, width: "6rem", flexShrink: 0 }}>Permit #</span>
-                          <span style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.ink }}>{job.permitNumber}</span>
-                        </div>
-                      )}
-                      {job.description && (
-                        <div style={{ display: "flex", gap: "0.75rem" }}>
-                          <span style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, width: "6rem", flexShrink: 0 }}>Description</span>
-                          <span style={{ fontSize: "0.8rem", color: S.ink, fontWeight: 300, lineHeight: 1.5 }}>{job.description}</span>
-                        </div>
-                      )}
-                      <div style={{ display: "flex", gap: "0.75rem" }}>
-                        <span style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, width: "6rem", flexShrink: 0 }}>Job ID</span>
-                        <span style={{ fontFamily: S.mono, fontSize: "0.6rem", color: S.inkLight }}>{job.id}</span>
-                      </div>
-                      {job.verified && (
-                        <div style={{ display: "flex", gap: "0.75rem", alignItems: "center" }}>
-                          <span style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, width: "6rem", flexShrink: 0 }}>ICP Record</span>
-                          <a
-                            href={`https://dashboard.internetcomputer.org/account/${property.owner}`}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            style={{ fontFamily: S.mono, fontSize: "0.6rem", color: S.sage, textDecoration: "none", borderBottom: `1px solid ${S.sage}` }}
-                          >
-                            Verified on ICP ↗
-                          </a>
-                        </div>
-                      )}
-                      {/* Edit button — only for homeowner-owned unverified jobs */}
-                      {!job.verified && job.homeowner === currentPrincipal && (
-                        <div style={{ marginTop: "0.25rem" }}>
-                          <button
-                            onClick={() => navigate("/jobs/new", { state: { editJob: job } })}
-                            style={{ fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.3rem 0.75rem", border: `1px solid ${S.rule}`, background: "#fff", color: S.inkLight, cursor: "pointer" }}
-                          >
-                            Edit record
-                          </button>
-                        </div>
-                      )}
-                      {/* Warranty doc upload — shown for any job with a warranty */}
-                      {job.warrantyMonths && job.warrantyMonths > 0 && (
-                        <div style={{ marginTop: "0.25rem", display: "flex", alignItems: "center", gap: "0.5rem" }}>
-                          <button
-                            onClick={() => warrantyInputRefs.current[job.id]?.click()}
-                            disabled={warrantyUploading === job.id}
-                            style={{ fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.3rem 0.75rem", border: `1px solid ${COLORS.sage}`, background: COLORS.white, color: COLORS.sage, cursor: warrantyUploading === job.id ? "not-allowed" : "pointer", opacity: warrantyUploading === job.id ? 0.5 : 1, display: "inline-flex", alignItems: "center", gap: "0.3rem" }}
-                          >
-                            🛡 {warrantyUploading === job.id ? "Uploading…" : "Upload warranty doc"}
-                          </button>
-                          <input
-                            ref={(el) => { warrantyInputRefs.current[job.id] = el; }}
-                            type="file"
-                            accept="image/*,application/pdf"
-                            style={{ display: "none" }}
-                            onChange={(e) => { const f = e.target.files?.[0]; if (f) handleWarrantyUpload(job, f); e.target.value = ""; }}
-                          />
-                        </div>
-                      )}
-                    </div>
-                  )}
-
-                  <PhotoStrip
-                    photos={photosByJob[job.id] ?? []}
-                    jobId={job.id}
-                    onUpload={onPhotoUpload}
-                  />
-                </div>
-              </div>
-            </React.Fragment>
-          );
-        })}
-      </div>
-    </>
-  );
-}
-
-function JobsTab({ jobs }: { jobs: Job[] }) {
-  const S = { rule: COLORS.rule, inkLight: COLORS.plumMid, ink: COLORS.plum, mono: FONTS.mono };
-
-  if (jobs.length === 0) {
-    return (
-      <div style={{ border: `1px dashed ${S.rule}`, padding: "3rem", textAlign: "center" }}>
-        <p style={{ fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.06em", color: S.inkLight }}>No jobs found.</p>
-      </div>
-    );
-  }
-
-  return (
-    <div style={{ border: `1px solid ${S.rule}` }}>
-      <table style={{ width: "100%", borderCollapse: "collapse" }}>
-        <thead>
-          <tr style={{ borderBottom: `1px solid ${S.rule}` }}>
-            {["Service", "Contractor", "Date", "Amount", "Status"].map((h) => (
-              <th key={h} style={{
-                textAlign: "left", padding: "0.75rem 1rem",
-                fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.12em", textTransform: "uppercase",
-                color: S.inkLight, fontWeight: 500,
-              }}>
-                {h}
-              </th>
-            ))}
-          </tr>
-        </thead>
-        <tbody>
-          {jobs.map((job, i) => (
-            <tr key={job.id} style={{ borderBottom: i < jobs.length - 1 ? `1px solid ${S.rule}` : "none", background: "#fff" }}>
-              <td style={{ padding: "0.875rem 1rem", fontWeight: 500, fontSize: "0.875rem" }}>{job.serviceType}</td>
-              <td style={{ padding: "0.875rem 1rem", fontSize: "0.875rem", color: S.inkLight }}>{job.isDiy ? "DIY" : job.contractorName}</td>
-              <td style={{ padding: "0.875rem 1rem", fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.06em", color: S.inkLight }}>{job.date}</td>
-              <td style={{ padding: "0.875rem 1rem", fontFamily: S.mono, fontSize: "0.875rem", fontWeight: 500 }}>${(job.amount / 100).toLocaleString()}</td>
-              <td style={{ padding: "0.875rem 1rem" }}>
-                <Badge variant={job.status === "verified" ? "success" : job.status === "completed" ? "info" : "warning"} size="sm">{job.status}</Badge>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-// ─── Document vault helpers ───────────────────────────────────────────────────
-
-const DOC_TYPES = ["Receipt", "Permit", "Inspection", "Warranty", "Invoice"] as const;
-type DocType = typeof DOC_TYPES[number];
-
-const DOC_TYPE_COLORS: Record<DocType, { color: string; bg: string }> = {
-  Receipt:    { color: COLORS.plumMid,  bg: COLORS.white },
-  Permit:     { color: COLORS.plum,     bg: COLORS.sageLight },
-  Inspection: { color: COLORS.sage,     bg: COLORS.sageLight },
-  Warranty:   { color: COLORS.plum,     bg: COLORS.butter },
-  Invoice:    { color: COLORS.sage,     bg: COLORS.blush },
-};
-
-function encodeDoc(type: DocType, filename: string): string {
-  return `[${type}] ${filename}`;
-}
-
-// Extended encoders for rich metadata types
-function encodePermit(permitNumber: string, authority: string, status: string, filename: string): string {
-  return `[Permit] ${permitNumber}|${authority}|${status}|${filename}`;
-}
-
-function encodeInspection(inspector: string, status: string, filename: string): string {
-  return `[Inspection] ${inspector}|${status}|${filename}`;
-}
-
-interface ParsedDoc {
-  type: DocType;
-  filename: string;
-  permitNumber?: string;
-  authority?: string;
-  inspector?: string;
-  status?: string;
-}
-
-function parseDoc(description: string): ParsedDoc {
-  const m = description.match(/^\[(\w+)\] (.+)$/);
-  if (!m || !DOC_TYPES.includes(m[1] as DocType)) {
-    return { type: "Receipt", filename: description };
-  }
-  const type = m[1] as DocType;
-  const rest = m[2];
-
-  if (type === "Permit") {
-    const parts = rest.split("|");
-    if (parts.length >= 4) {
-      return { type, permitNumber: parts[0], authority: parts[1], status: parts[2], filename: parts.slice(3).join("|") };
-    }
-  }
-  if (type === "Inspection") {
-    const parts = rest.split("|");
-    if (parts.length >= 3) {
-      return { type, inspector: parts[0], status: parts[1], filename: parts.slice(2).join("|") };
-    }
-  }
-  return { type, filename: rest };
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-
-type BatchFileStatus = "pending" | "uploading" | "done" | "duplicate" | "error";
-interface BatchFile { name: string; status: BatchFileStatus; error?: string }
-
-function DocumentsTab({ propertyId }: { propertyId: string }) {
-  const S = { ink: COLORS.plum, rule: COLORS.rule, inkLight: COLORS.plumMid, rust: COLORS.sage, serif: FONTS.serif, mono: FONTS.mono };
-  const DOCS_JOB = `docs_${propertyId}`;
-  const inputRef  = React.useRef<HTMLInputElement>(null);
-  const permitRef = React.useRef<HTMLInputElement>(null);
-  const inspectionRef = React.useRef<HTMLInputElement>(null);
-
-  const [docs,     setDocs]     = useState<Photo[]>([]);
-  const [docType,  setDocType]  = useState<DocType>("Receipt");
-  const [queue,    setQueue]    = useState<BatchFile[]>([]);
-  const batchActive = queue.some((f) => f.status === "pending" || f.status === "uploading");
-
-  // Permit metadata
-  const [permitNumber,    setPermitNumber]    = useState("");
-  const [permitAuthority, setPermitAuthority] = useState("");
-  const [permitStatus,    setPermitStatus]    = useState<"Open" | "Closed" | "Expired">("Open");
-  const [permitUploading, setPermitUploading] = useState(false);
-
-  // Inspection metadata
-  const [inspectorName,     setInspectorName]     = useState("");
-  const [inspectionStatus,  setInspectionStatus]  = useState<"Pass" | "Fail" | "Conditional">("Pass");
-  const [inspectionUploading, setInspectionUploading] = useState(false);
-
-  const handlePermitUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    if (permitRef.current) permitRef.current.value = "";
-    setPermitUploading(true);
-    try {
-      const description = encodePermit(permitNumber || "No #", permitAuthority || "Unknown", permitStatus, file.name);
-      const doc = await photoService.upload(file, DOCS_JOB, propertyId, "PostConstruction", description);
-      setDocs((prev) => [doc, ...prev]);
-      toast.success("Permit uploaded");
-      setPermitNumber(""); setPermitAuthority("");
-    } catch (err: any) {
-      const msg: string = err.message ?? "Upload failed";
-      toast.error(msg === "Duplicate" ? "Already uploaded" : msg);
-    } finally {
-      setPermitUploading(false);
-    }
-  };
-
-  const handleInspectionUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    if (inspectionRef.current) inspectionRef.current.value = "";
-    setInspectionUploading(true);
-    try {
-      const description = encodeInspection(inspectorName || "Unknown", inspectionStatus, file.name);
-      const doc = await photoService.upload(file, DOCS_JOB, propertyId, "PostConstruction", description);
-      setDocs((prev) => [doc, ...prev]);
-      toast.success("Inspection report uploaded");
-      setInspectorName("");
-    } catch (err: any) {
-      const msg: string = err.message ?? "Upload failed";
-      toast.error(msg === "Duplicate" ? "Already uploaded" : msg);
-    } finally {
-      setInspectionUploading(false);
-    }
-  };
-
-  useEffect(() => {
-    // Load both legacy "receipts_" key and new "docs_" key
-    Promise.all([
-      photoService.getByJob(`receipts_${propertyId}`).catch(() => []),
-      photoService.getByJob(DOCS_JOB).catch(() => []),
-    ]).then(([legacy, current]) => {
-      const all = [...legacy, ...current].sort((a, b) => b.createdAt - a.createdAt);
-      setDocs(all);
-    });
-  }, [propertyId]);
-
-  const handleUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const files = Array.from(e.target.files ?? []);
-    if (files.length === 0) return;
-    if (inputRef.current) inputRef.current.value = "";
-
-    const initialQueue: BatchFile[] = files.map((f) => ({ name: f.name, status: "pending" }));
-    setQueue(initialQueue);
-
-    for (let i = 0; i < files.length; i++) {
-      const file = files[i];
-      setQueue((prev) => prev.map((q, idx) => idx === i ? { ...q, status: "uploading" } : q));
-      try {
-        const description = encodeDoc(docType, file.name);
-        const doc = await photoService.upload(file, DOCS_JOB, propertyId, "PostConstruction", description);
-        setDocs((prev) => [doc, ...prev]);
-        setQueue((prev) => prev.map((q, idx) => idx === i ? { ...q, status: "done" } : q));
-      } catch (err: any) {
-        const msg: string = err.message ?? "Upload failed";
-        const isDuplicate = msg === "Duplicate" || msg.startsWith("Duplicate");
-        setQueue((prev) => prev.map((q, idx) => idx === i ? { ...q, status: isDuplicate ? "duplicate" : "error", error: isDuplicate ? "Already uploaded" : msg } : q));
-      }
-    }
-  };
-
-  const statusIcon = (s: BatchFileStatus) => {
-    if (s === "done")      return <span style={{ color: COLORS.sage }}>✓</span>;
-    if (s === "duplicate") return <span style={{ color: COLORS.plumMid }}>⊘</span>;
-    if (s === "error")     return <span style={{ color: COLORS.plum }}>✗</span>;
-    if (s === "uploading") return <span style={{ color: COLORS.plum }}>↑</span>;
-    return <span style={{ color: COLORS.rule }}>…</span>;
-  };
-
-  return (
-    <div>
-      {/* Permits & Inspections — dedicated section */}
-      <div style={{ border: `1px solid ${COLORS.rule}`, borderRadius: RADIUS.sm, marginBottom: "1.25rem", overflow: "hidden" }}>
-        <div style={{ padding: "0.875rem 1.25rem", borderBottom: `1px solid ${COLORS.rule}`, background: COLORS.sageLight, display: "flex", justifyContent: "space-between", alignItems: "center" }}>
-          <p style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.12em", textTransform: "uppercase", color: COLORS.plum }}>
-            Permits &amp; Inspections
-          </p>
-          <p style={{ fontFamily: S.mono, fontSize: "0.55rem", color: COLORS.plumMid }}>Upload with metadata — status tracked on-chain</p>
-        </div>
-        <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", background: COLORS.white }}>
-          {/* Permit */}
-          <div style={{ padding: "1rem 1.25rem", borderRight: `1px solid ${S.rule}` }}>
-            <p style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", color: COLORS.plum, marginBottom: "0.75rem" }}>Permit</p>
-            <div style={{ display: "flex", flexDirection: "column", gap: "0.5rem", marginBottom: "0.75rem" }}>
-              <input
-                className="form-input"
-                placeholder="Permit #"
-                value={permitNumber}
-                onChange={(e) => setPermitNumber(e.target.value)}
-                style={{ fontSize: "0.8rem" }}
-              />
-              <input
-                className="form-input"
-                placeholder="Issuing authority (e.g. City of Austin)"
-                value={permitAuthority}
-                onChange={(e) => setPermitAuthority(e.target.value)}
-                style={{ fontSize: "0.8rem" }}
-              />
-              <select
-                className="form-input"
-                value={permitStatus}
-                onChange={(e) => setPermitStatus(e.target.value as any)}
-                style={{ fontSize: "0.8rem" }}
-              >
-                <option value="Open">Open</option>
-                <option value="Closed">Closed</option>
-                <option value="Expired">Expired</option>
-              </select>
-            </div>
-            <button
-              disabled={permitUploading}
-              onClick={() => permitRef.current?.click()}
-              style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.375rem 0.875rem", border: `1px solid ${COLORS.plum}`, color: COLORS.plum, background: "none", cursor: permitUploading ? "not-allowed" : "pointer", opacity: permitUploading ? 0.5 : 1 }}
-            >
-              {permitUploading ? "Uploading…" : "+ Upload Permit"}
-            </button>
-            <input ref={permitRef} type="file" accept="image/*,application/pdf" style={{ display: "none" }} onChange={handlePermitUpload} />
-          </div>
-          {/* Inspection */}
-          <div style={{ padding: "1rem 1.25rem" }}>
-            <p style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", color: COLORS.sage, marginBottom: "0.75rem" }}>Inspection Report</p>
-            <div style={{ display: "flex", flexDirection: "column", gap: "0.5rem", marginBottom: "0.75rem" }}>
-              <input
-                className="form-input"
-                placeholder="Inspector name or company"
-                value={inspectorName}
-                onChange={(e) => setInspectorName(e.target.value)}
-                style={{ fontSize: "0.8rem" }}
-              />
-              <select
-                className="form-input"
-                value={inspectionStatus}
-                onChange={(e) => setInspectionStatus(e.target.value as any)}
-                style={{ fontSize: "0.8rem" }}
-              >
-                <option value="Pass">Pass</option>
-                <option value="Conditional">Conditional</option>
-                <option value="Fail">Fail</option>
-              </select>
-            </div>
-            <button
-              disabled={inspectionUploading}
-              onClick={() => inspectionRef.current?.click()}
-              style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.375rem 0.875rem", border: `1px solid ${COLORS.sage}`, color: COLORS.sage, background: "none", cursor: inspectionUploading ? "not-allowed" : "pointer", opacity: inspectionUploading ? 0.5 : 1 }}
-            >
-              {inspectionUploading ? "Uploading…" : "+ Upload Report"}
-            </button>
-            <input ref={inspectionRef} type="file" accept="image/*,application/pdf" style={{ display: "none" }} onChange={handleInspectionUpload} />
-          </div>
-        </div>
-      </div>
-
-      {/* Upload controls */}
-      <div style={{ border: `1px solid ${S.rule}`, marginBottom: "1.5rem" }}>
-        <div style={{ padding: "0.875rem 1.25rem", borderBottom: `1px solid ${S.rule}`, display: "flex", alignItems: "center", justifyContent: "space-between" }}>
-          <p style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.12em", textTransform: "uppercase", color: S.inkLight }}>
-            Upload Documents
-          </p>
-          <p style={{ fontFamily: S.mono, fontSize: "0.55rem", color: S.inkLight }}>
-            Select multiple files — duplicates are auto-detected
-          </p>
-        </div>
-        <div style={{ padding: "1rem 1.25rem", display: "flex", alignItems: "center", gap: "0.75rem", flexWrap: "wrap" }}>
-          {/* Type picker */}
-          <div style={{ display: "flex", gap: "1px", background: S.rule }}>
-            {DOC_TYPES.map((t) => (
-              <button
-                key={t}
-                onClick={() => setDocType(t)}
-                style={{
-                  padding: "0.35rem 0.75rem",
-                  fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase",
-                  border: "none", cursor: "pointer",
-                  background: docType === t ? COLORS.plum : COLORS.white,
-                  color:      docType === t ? COLORS.white : S.inkLight,
-                }}
-              >
-                {t}
-              </button>
-            ))}
-          </div>
-          {/* Upload button */}
-          <button
-            onClick={() => inputRef.current?.click()}
-            disabled={batchActive}
-            style={{
-              fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.1em", textTransform: "uppercase",
-              padding: "0.375rem 0.875rem", border: `1px solid ${S.rust}`, color: S.rust,
-              background: "none", cursor: batchActive ? "not-allowed" : "pointer", opacity: batchActive ? 0.5 : 1,
-            }}
-          >
-            {batchActive ? "Uploading…" : `+ Upload ${docType}s`}
-          </button>
-          <input ref={inputRef} type="file" multiple accept="image/*,application/pdf" style={{ display: "none" }} onChange={handleUpload} />
-        </div>
-
-        {/* Batch progress queue */}
-        {queue.length > 0 && (
-          <div style={{ borderTop: `1px solid ${S.rule}` }}>
-            {queue.map((f, i) => (
-              <div key={i} style={{ display: "flex", alignItems: "center", gap: "0.75rem", padding: "0.5rem 1.25rem", borderBottom: i < queue.length - 1 ? `1px solid ${S.rule}` : "none", background: f.status === "error" ? COLORS.blush : f.status === "duplicate" ? COLORS.sageLight : COLORS.white }}>
-                <span style={{ fontFamily: S.mono, fontSize: "0.8rem", width: "1rem", textAlign: "center" }}>{statusIcon(f.status)}</span>
-                <span style={{ flex: 1, fontFamily: S.mono, fontSize: "0.65rem", overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>{f.name}</span>
-                <span style={{ fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.06em", textTransform: "uppercase", color: f.status === "error" ? S.rust : f.status === "duplicate" ? S.inkLight : f.status === "done" ? COLORS.sage : S.inkLight }}>
-                  {f.status === "error" ? (f.error ?? "Error") : f.status === "duplicate" ? "Duplicate — skipped" : f.status === "done" ? "Uploaded" : f.status === "uploading" ? "Uploading…" : "Queued"}
-                </span>
-              </div>
-            ))}
-            {!batchActive && (
-              <div style={{ padding: "0.5rem 1.25rem", display: "flex", justifyContent: "flex-end" }}>
-                <button onClick={() => setQueue([])} style={{ fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase", border: "none", background: "none", color: S.inkLight, cursor: "pointer" }}>
-                  Clear
-                </button>
-              </div>
-            )}
-          </div>
-        )}
-      </div>
-
-      {/* Document list */}
-      {docs.length === 0 ? (
-        <div style={{ border: `1px dashed ${S.rule}`, padding: "3rem", textAlign: "center" }}>
-          <p style={{ fontFamily: S.serif, fontWeight: 700, marginBottom: "0.375rem" }}>No documents yet</p>
-          <p style={{ fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.06em", color: S.inkLight }}>
-            Upload receipts, permits, inspection reports, warranties, or invoices.
-            Each file is SHA-256 hashed and stored on-chain.
-          </p>
-        </div>
-      ) : (
-        <div style={{ border: `1px solid ${S.rule}` }}>
-          {docs.map((doc, i) => {
-            const parsed = parseDoc(doc.description);
-            const { type, filename } = parsed;
-            const tc = DOC_TYPE_COLORS[type];
-            return (
-              <div key={doc.id} style={{
-                display: "flex", alignItems: "center", gap: "1rem", padding: "0.875rem 1.25rem",
-                background: "#fff", borderBottom: i < docs.length - 1 ? `1px solid ${S.rule}` : "none",
-              }}>
-                {/* Type badge */}
-                <span style={{
-                  fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.1em", textTransform: "uppercase",
-                  padding: "0.2rem 0.5rem", flexShrink: 0,
-                  color: tc.color, background: tc.bg, border: `1px solid ${tc.color}30`,
-                }}>
-                  {type}
-                </span>
-                <div style={{ flex: 1, minWidth: 0 }}>
-                  <p style={{ fontSize: "0.875rem", fontWeight: 500, marginBottom: "0.125rem", overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>
-                    {filename}
-                  </p>
-                  {/* Permit metadata */}
-                  {type === "Permit" && parsed.permitNumber && (
-                    <p style={{ fontFamily: S.mono, fontSize: "0.55rem", color: COLORS.plum, letterSpacing: "0.06em", marginBottom: "0.1rem" }}>
-                      {parsed.permitNumber} · {parsed.authority} · <span style={{ textTransform: "uppercase", fontWeight: 700, color: parsed.status === "Closed" ? COLORS.sage : parsed.status === "Expired" ? COLORS.plumMid : COLORS.plum }}>{parsed.status}</span>
-                    </p>
-                  )}
-                  {/* Inspection metadata */}
-                  {type === "Inspection" && parsed.inspector && (
-                    <p style={{ fontFamily: S.mono, fontSize: "0.55rem", color: COLORS.sage, letterSpacing: "0.06em", marginBottom: "0.1rem" }}>
-                      {parsed.inspector} · <span style={{ textTransform: "uppercase", fontWeight: 700, color: parsed.status === "Pass" ? COLORS.sage : parsed.status === "Fail" ? COLORS.plum : COLORS.plumMid }}>{parsed.status}</span>
-                    </p>
-                  )}
-                  <p style={{ fontFamily: S.mono, fontSize: "0.6rem", color: S.inkLight, letterSpacing: "0.06em" }}>
-                    {(doc.size / 1024).toFixed(1)} KB · {doc.hash.slice(0, 16)}… · {new Date(doc.createdAt).toLocaleDateString()}
-                  </p>
-                </div>
-                {doc.url && (
-                  <a href={doc.url} target="_blank" rel="noreferrer" style={{
-                    fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase",
-                    color: S.rust, textDecoration: "none", flexShrink: 0,
-                  }}>
-                    View
-                  </a>
-                )}
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function SettingsTab({ property, currentPrincipal }: { property: Property; currentPrincipal: string }) {
-  const S = { rule: COLORS.rule, inkLight: COLORS.plumMid, ink: COLORS.plum, rust: COLORS.sage, sage: COLORS.sage, paper: COLORS.white, serif: FONTS.serif, mono: FONTS.mono };
-  const navigate = useNavigate();
-
-  const [transferStep, setTransferStep] = React.useState<"idle" | "loading" | "done">("idle");
-  const [transferToken, setTransferToken]   = React.useState<string | null>(null);
-  const [transferExpiry, setTransferExpiry] = React.useState<Date | null>(null);
-  const [transferError, setTransferError]   = React.useState<string | null>(null);
-  const [copied, setCopied]                 = React.useState(false);
-  const [cancelLoading, setCancelLoading]   = React.useState(false);
-
-  const [historyRecords, setHistoryRecords] = React.useState<import("../services/property").TransferRecord[]>([]);
-
-  // Manager delegation state
-  const [managers, setManagers]               = React.useState<import("../services/property").PropertyManager[]>([]);
-  const [removingManager, setRemovingManager] = React.useState<string | null>(null);
-  const [inviteDisplayName, setInviteDisplayName] = React.useState("");
-  const [inviteRole, setInviteRole]           = React.useState<import("../services/property").ManagerRole>("Viewer");
-  const [inviteStep, setInviteStep]           = React.useState<"idle" | "loading" | "done">("idle");
-  const [inviteToken, setInviteToken]         = React.useState<string | null>(null);
-  const [inviteExpiry, setInviteExpiry]       = React.useState<Date | null>(null);
-  const [inviteError, setInviteError]         = React.useState<string | null>(null);
-  const [inviteCopied, setInviteCopied]       = React.useState(false);
-
-  // Reload any existing pending transfer for this property (e.g. after page refresh)
-  React.useEffect(() => {
-    propertyService.getPendingTransfer(BigInt(property.id)).then((pt) => {
-      if (pt && pt.from === currentPrincipal) {
-        setTransferToken(pt.token);
-        setTransferExpiry(new Date(pt.expiresAt));
-        setTransferStep("done");
-      }
-    }).catch(() => {});
-    propertyService.getOwnershipHistory(BigInt(property.id)).then(setHistoryRecords).catch(() => {});
-    propertyService.getPropertyManagers(BigInt(property.id)).then(setManagers).catch(() => {});
-  }, [property.id, currentPrincipal]);
-
-  const inviteUrl = inviteToken
-    ? `${window.location.origin}/manage/claim/${inviteToken}`
-    : null;
-
-  const claimUrl = transferToken
-    ? `${window.location.origin}/transfer/claim/${transferToken}`
-    : null;
-
-  const verificationNext =
-    property.verificationLevel === "Unverified"
-      ? { label: "Verify Ownership", href: `/properties/${property.id}/verify`, color: S.rust }
-      : property.verificationLevel === "Basic"
-      ? { label: "Upgrade to Premium", href: "/pricing", color: S.ink }
-      : null;
-
-  const section = (title: string) => (
-    <div style={{ padding: "0.875rem 1.25rem", borderBottom: `1px solid ${S.rule}`, background: S.paper }}>
-      <p style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.12em", textTransform: "uppercase", color: S.inkLight }}>{title}</p>
-    </div>
-  );
-
-  return (
-    <div style={{ display: "flex", flexDirection: "column", gap: "1.25rem" }}>
-
-      {/* Property Details */}
-      <div style={{ border: `1px solid ${S.rule}` }}>
-        {section("Property Details")}
-        {[
-          { label: "Address",     value: property.address },
-          { label: "City",        value: property.city },
-          { label: "State",       value: property.state },
-          { label: "ZIP Code",    value: property.zipCode },
-          { label: "Type",        value: property.propertyType },
-          { label: "Year Built",  value: String(property.yearBuilt) },
-          { label: "Square Feet", value: `${Number(property.squareFeet).toLocaleString()} sq ft` },
-        ].map((row, i, arr) => (
-          <div key={row.label} style={{ display: "flex", justifyContent: "space-between", alignItems: "center", padding: "0.75rem 1.25rem", borderBottom: i < arr.length - 1 ? `1px solid ${S.rule}` : "none", background: "#fff" }}>
-            <span style={{ fontFamily: S.mono, fontSize: "0.65rem", letterSpacing: "0.1em", textTransform: "uppercase", color: S.inkLight }}>{row.label}</span>
-            <span style={{ fontSize: "0.875rem", fontWeight: 500, color: S.ink }}>{row.value}</span>
-          </div>
-        ))}
-      </div>
-
-      {/* Verification & Trust */}
-      <div style={{ border: `1px solid ${S.rule}` }}>
-        {section("Verification & Trust")}
-        <div style={{ padding: "1.25rem", background: "#fff", display: "flex", alignItems: "flex-start", justifyContent: "space-between", flexWrap: "wrap", gap: "1rem" }}>
-          <div style={{ flex: 1, minWidth: 0 }}>
-            <p style={{ fontFamily: S.mono, fontSize: "0.65rem", fontWeight: 700, letterSpacing: "0.08em", textTransform: "uppercase", marginBottom: "0.25rem",
-              color: property.verificationLevel === "Premium" ? S.sage : property.verificationLevel === "Basic" ? COLORS.plum : property.verificationLevel === "PendingReview" ? COLORS.plumMid : S.inkLight }}>
-              {property.verificationLevel}
-            </p>
-            <p style={{ fontSize: "0.8rem", color: S.inkLight, fontWeight: 300, lineHeight: 1.5 }}>
-              {property.verificationLevel === "Premium"
-                ? "Fully verified — buyers and lenders trust this record."
-                : property.verificationLevel === "Basic"
-                ? "Basic verification complete. Upgrade for full buyer trust."
-                : property.verificationLevel === "PendingReview"
-                ? "Your documents are under review. We'll notify you when done."
-                : "Not yet verified. Upload ownership documents to unlock report sharing."}
-            </p>
-          </div>
-          {verificationNext && (
-            <button
-              onClick={() => navigate(verificationNext.href)}
-              style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.5rem 1rem", background: verificationNext.color, color: "#fff", border: "none", cursor: "pointer", flexShrink: 0 }}
-            >
-              {verificationNext.label} →
-            </button>
-          )}
-        </div>
-      </div>
-
-      {/* Plan & Limits */}
-      <div style={{ border: `1px solid ${S.rule}` }}>
-        {section("Plan & Limits")}
-        <div style={{ padding: "1.25rem", background: "#fff", display: "flex", alignItems: "flex-start", justifyContent: "space-between", flexWrap: "wrap", gap: "1rem" }}>
-          <div style={{ flex: 1, minWidth: 0 }}>
-            <p style={{ fontFamily: S.serif, fontWeight: 900, fontSize: "1.1rem", color: S.ink, marginBottom: "0.25rem" }}>
-              {property.tier ?? "Free"}
-            </p>
-            <p style={{ fontSize: "0.8rem", color: S.inkLight, fontWeight: 300, lineHeight: 1.5 }}>
-              {property.tier === "Pro"
-                ? "5 properties · 10 photos per job · 10 open quotes"
-                : property.tier === "Premium"
-                ? "20 properties · 30 photos per job · full history exports"
-                : "1 property · 2 photos per job · 3 open quotes"}
-            </p>
-          </div>
-          {(!property.tier || property.tier === "Free") && (
-            <button
-              onClick={() => navigate("/pricing")}
-              style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.5rem 1rem", background: S.ink, color: S.paper, border: "none", cursor: "pointer", flexShrink: 0 }}
-            >
-              Upgrade Plan →
-            </button>
-          )}
-        </div>
-      </div>
-
-      {/* On-Chain Identity */}
-      <div style={{ border: `1px solid ${S.rule}` }}>
-        {section("On-Chain Identity")}
-        <div style={{ padding: "1.25rem", background: "#fff", display: "flex", flexDirection: "column", gap: "0.75rem" }}>
-          {[
-            { label: "Owner Principal", value: property.owner },
-            { label: "Property ID",     value: String(property.id) },
-          ].map((row) => (
-            <div key={row.label} style={{ display: "flex", gap: "0.75rem", alignItems: "flex-start" }}>
-              <span style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, width: "8rem", flexShrink: 0, paddingTop: "0.1rem" }}>{row.label}</span>
-              <span style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.ink, wordBreak: "break-all" }}>{row.value}</span>
-            </div>
-          ))}
-          <div style={{ display: "flex", gap: "0.75rem", alignItems: "center" }}>
-            <span style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, width: "8rem", flexShrink: 0 }}>ICP Dashboard</span>
-            <a
-              href={`https://dashboard.internetcomputer.org/account/${property.owner}`}
-              target="_blank"
-              rel="noopener noreferrer"
-              style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.sage, textDecoration: "none", borderBottom: `1px solid ${S.sage}` }}
-            >
-              View on ICP Explorer ↗
-            </a>
-          </div>
-        </div>
-      </div>
-
-      {/* Transfer Ownership */}
-      <div style={{ border: `1px solid ${S.rust}` }}>
-        {section("Transfer Ownership")}
-        <div style={{ padding: "1.25rem", background: "#fff", display: "flex", flexDirection: "column", gap: "0.875rem" }}>
-
-          {transferStep === "idle" && (
-            <>
-              <p style={{ fontSize: "0.8rem", color: S.inkLight, fontWeight: 300, lineHeight: 1.6 }}>
-                Generate a secure link and share it with the buyer. They'll use it to claim this property — all history, photos, and maintenance records transfer to them automatically.
-              </p>
-              <p style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.inkLight, lineHeight: 1.5 }}>
-                The link expires in <strong>90 days</strong>. Ownership only transfers when the buyer claims it — generating the link doesn't move anything yet.
-              </p>
-              {transferError && (
-                <p style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.rust }}>{transferError}</p>
-              )}
-              <button
-                onClick={async () => {
-                  setTransferStep("loading");
-                  setTransferError(null);
-                  try {
-                    const pt = await propertyService.initiateTransfer(BigInt(property.id));
-                    setTransferToken(pt.token);
-                    setTransferExpiry(new Date(pt.expiresAt));
-                    setTransferStep("done");
-                  } catch (e: any) {
-                    setTransferError(e.message ?? "Failed to generate transfer link.");
-                    setTransferStep("idle");
-                  }
-                }}
-                style={{ alignSelf: "flex-start", fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.5rem 1rem", background: S.rust, color: "#fff", border: "none", cursor: "pointer" }}
-              >
-                Generate Transfer Link →
-              </button>
-            </>
-          )}
-
-          {transferStep === "loading" && (
-            <p style={{ fontFamily: S.mono, fontSize: "0.7rem", color: S.inkLight }}>Generating link…</p>
-          )}
-
-          {transferStep === "done" && claimUrl && (
-            <>
-              <p style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.inkLight, lineHeight: 1.5 }}>
-                Share this link with the buyer. Ownership transfers the moment they log in and accept.
-              </p>
-
-              {/* Link copy box */}
-              <div style={{ display: "flex", gap: "0.5rem", alignItems: "stretch" }}>
-                <div style={{
-                  flex: 1, padding: "0.6rem 0.75rem",
-                  border: `1px solid ${S.rule}`, background: COLORS.sageLight,
-                  fontFamily: S.mono, fontSize: "0.6rem", color: S.ink,
-                  wordBreak: "break-all", lineHeight: 1.5,
-                }}>
-                  {claimUrl}
-                </div>
-                <button
-                  onClick={() => {
-                    navigator.clipboard.writeText(claimUrl).then(() => {
-                      setCopied(true);
-                      setTimeout(() => setCopied(false), 2000);
-                    });
-                  }}
-                  style={{
-                    fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase",
-                    padding: "0 0.875rem", background: copied ? S.sage : S.ink,
-                    color: "#fff", border: "none", cursor: "pointer", flexShrink: 0, transition: "background 0.15s",
-                  }}
-                >
-                  {copied ? "Copied!" : "Copy"}
-                </button>
-              </div>
-
-              {/* Expiry */}
-              {transferExpiry && (
-                <p style={{ fontFamily: S.mono, fontSize: "0.6rem", color: S.inkLight }}>
-                  Expires {transferExpiry.toLocaleDateString(undefined, { year: "numeric", month: "long", day: "numeric" })}
-                </p>
-              )}
-
-              {/* Cancel */}
-              <button
-                disabled={cancelLoading}
-                onClick={async () => {
-                  setCancelLoading(true);
-                  try {
-                    await propertyService.cancelTransfer(BigInt(property.id));
-                    setTransferToken(null);
-                    setTransferExpiry(null);
-                    setTransferStep("idle");
-                    toast.success("Transfer link cancelled.");
-                  } catch (e: any) {
-                    toast.error(e.message ?? "Could not cancel transfer.");
-                  } finally {
-                    setCancelLoading(false);
-                  }
-                }}
-                style={{ alignSelf: "flex-start", fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", padding: "0.375rem 0.75rem", background: "none", border: `1px solid ${S.rule}`, color: S.inkLight, cursor: "pointer" }}
-              >
-                {cancelLoading ? "Cancelling…" : "Cancel Transfer"}
-              </button>
-            </>
-          )}
-
-        </div>
-      </div>
-
-      {/* Ownership History */}
-      {historyRecords.length > 0 && (
-        <div style={{ border: `1px solid ${S.rule}` }}>
-          {section("Ownership History")}
-          <div style={{ background: "#fff" }}>
-            {historyRecords.map((r, i) => (
-              <div
-                key={i}
-                style={{ padding: "0.875rem 1.25rem", borderBottom: i < historyRecords.length - 1 ? `1px solid ${S.rule}` : "none", display: "flex", flexDirection: "column", gap: "0.25rem" }}
-              >
-                <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center" }}>
-                  <span style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight }}>
-                    {new Date(r.timestamp).toLocaleDateString()}
-                  </span>
-                  {r.txHash && (
-                    <span style={{ fontFamily: S.mono, fontSize: "0.55rem", color: S.inkLight, opacity: 0.7 }}>{r.txHash.slice(0, 16)}…</span>
-                  )}
-                </div>
-                <div style={{ fontFamily: S.mono, fontSize: "0.6rem", color: S.ink }}>
-                  <span style={{ color: S.inkLight }}>From </span>{r.from.slice(0, 20)}…
-                  <span style={{ color: S.inkLight }}> → </span>{r.to.slice(0, 20)}…
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* Access & Managers */}
-      <div style={{ border: `1px solid ${S.rule}` }}>
-        {section("Access & Managers")}
-        <div style={{ padding: "1.25rem", background: "#fff", display: "flex", flexDirection: "column", gap: "1rem" }}>
-
-          <p style={{ fontSize: "0.8rem", color: S.inkLight, fontWeight: 300, lineHeight: 1.6, margin: 0 }}>
-            Grant a family member or property manager access to this property. <strong>Viewer</strong> can see all records. <strong>Manager</strong> can also add jobs, photos, and maintenance entries.
-          </p>
-
-          {/* Current managers list */}
-          {managers.length > 0 && (
-            <div style={{ border: `1px solid ${S.rule}` }}>
-              {managers.map((m, i) => (
-                <div
-                  key={m.principal}
-                  style={{ display: "flex", alignItems: "center", justifyContent: "space-between", padding: "0.75rem 1rem", borderBottom: i < managers.length - 1 ? `1px solid ${S.rule}` : "none", gap: "0.75rem" }}
-                >
-                  <div style={{ flex: 1, minWidth: 0 }}>
-                    <p style={{ fontFamily: S.mono, fontSize: "0.7rem", fontWeight: 600, color: S.ink, margin: 0, marginBottom: "0.15rem" }}>{m.displayName}</p>
-                    <p style={{ fontFamily: S.mono, fontSize: "0.55rem", color: S.inkLight, margin: 0, wordBreak: "break-all" }}>{m.principal}</p>
-                  </div>
-                  <span style={{
-                    fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase",
-                    padding: "0.2rem 0.5rem", border: `1px solid ${m.role === "Manager" ? S.ink : S.rule}`,
-                    color: m.role === "Manager" ? S.ink : S.inkLight, flexShrink: 0,
-                  }}>
-                    {m.role}
-                  </span>
-                  <button
-                    disabled={removingManager === m.principal}
-                    onClick={async () => {
-                      setRemovingManager(m.principal);
-                      try {
-                        await propertyService.removeManager(BigInt(property.id), m.principal);
-                        setManagers((prev) => prev.filter((x) => x.principal !== m.principal));
-                      } catch (e: any) {
-                        toast.error(e.message ?? "Could not remove manager.");
-                      } finally {
-                        setRemovingManager(null);
-                      }
-                    }}
-                    style={{ fontFamily: S.mono, fontSize: "0.55rem", letterSpacing: "0.08em", textTransform: "uppercase", padding: "0.2rem 0.5rem", background: "none", border: `1px solid ${S.rule}`, color: S.inkLight, cursor: "pointer", flexShrink: 0 }}
-                  >
-                    {removingManager === m.principal ? "Removing…" : "Remove"}
-                  </button>
-                </div>
-              ))}
-            </div>
-          )}
-
-          {/* Invite form */}
-          {inviteStep === "idle" && (
-            <div style={{ display: "flex", flexDirection: "column", gap: "0.75rem" }}>
-              <div style={{ display: "flex", gap: "0.75rem", alignItems: "flex-end" }}>
-                <div style={{ flex: 1 }}>
-                  <label style={{ display: "block", fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight, marginBottom: "0.35rem" }}>
-                    Display Name
-                  </label>
-                  <input
-                    type="text"
-                    value={inviteDisplayName}
-                    onChange={(e) => setInviteDisplayName(e.target.value)}
-                    placeholder="e.g. Sarah (daughter)"
-                    style={{ width: "100%", fontFamily: S.mono, fontSize: "0.7rem", padding: "0.5rem 0.6rem", border: `1px solid ${S.rule}`, background: "#fff", color: S.ink, outline: "none", boxSizing: "border-box" }}
-                  />
-                </div>
-                <div style={{ display: "flex", flexDirection: "column", gap: "0.35rem" }}>
-                  <label style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: S.inkLight }}>Role</label>
-                  <div style={{ display: "flex", gap: "0.5rem" }}>
-                    {(["Viewer", "Manager"] as const).map((r) => (
-                      <button
-                        key={r}
-                        onClick={() => setInviteRole(r)}
-                        style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", padding: "0.4rem 0.75rem", background: inviteRole === r ? S.ink : "none", color: inviteRole === r ? "#fff" : S.inkLight, border: `1px solid ${inviteRole === r ? S.ink : S.rule}`, cursor: "pointer" }}
-                      >
-                        {r}
-                      </button>
-                    ))}
-                  </div>
-                </div>
-              </div>
-              {inviteError && (
-                <p style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.rust, margin: 0 }}>{inviteError}</p>
-              )}
-              <button
-                disabled={!inviteDisplayName.trim()}
-                onClick={async () => {
-                  setInviteStep("loading");
-                  setInviteError(null);
-                  try {
-                    const invite = await propertyService.inviteManager(BigInt(property.id), inviteRole, inviteDisplayName.trim());
-                    setInviteToken(invite.token);
-                    setInviteExpiry(new Date(invite.expiresAt));
-                    setInviteStep("done");
-                  } catch (e: any) {
-                    setInviteError(e.message ?? "Failed to generate invite link.");
-                    setInviteStep("idle");
-                  }
-                }}
-                style={{ alignSelf: "flex-start", fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", padding: "0.5rem 1rem", background: inviteDisplayName.trim() ? S.ink : S.rule, color: "#fff", border: "none", cursor: inviteDisplayName.trim() ? "pointer" : "default" }}
-              >
-                Generate Invite Link →
-              </button>
-            </div>
-          )}
-
-          {inviteStep === "loading" && (
-            <p style={{ fontFamily: S.mono, fontSize: "0.7rem", color: S.inkLight, margin: 0 }}>Generating invite…</p>
-          )}
-
-          {inviteStep === "done" && inviteUrl && (
-            <div style={{ display: "flex", flexDirection: "column", gap: "0.75rem" }}>
-              <p style={{ fontFamily: S.mono, fontSize: "0.65rem", color: S.inkLight, margin: 0, lineHeight: 1.5 }}>
-                Share this link with <strong>{inviteDisplayName}</strong>. They'll log in and accept the <strong>{inviteRole}</strong> role. The link expires in 90 days.
-              </p>
-              <div style={{ display: "flex", gap: "0.5rem", alignItems: "stretch" }}>
-                <div style={{ flex: 1, padding: "0.6rem 0.75rem", border: `1px solid ${S.rule}`, background: COLORS.sageLight, fontFamily: S.mono, fontSize: "0.6rem", color: S.ink, wordBreak: "break-all", lineHeight: 1.5 }}>
-                  {inviteUrl}
-                </div>
-                <button
-                  onClick={() => {
-                    navigator.clipboard.writeText(inviteUrl).then(() => {
-                      setInviteCopied(true);
-                      setTimeout(() => setInviteCopied(false), 2000);
-                    });
-                  }}
-                  style={{ fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", padding: "0 0.875rem", background: inviteCopied ? S.sage : S.ink, color: "#fff", border: "none", cursor: "pointer", flexShrink: 0, transition: "background 0.15s" }}
-                >
-                  {inviteCopied ? "Copied!" : "Copy"}
-                </button>
-              </div>
-              {inviteExpiry && (
-                <p style={{ fontFamily: S.mono, fontSize: "0.6rem", color: S.inkLight, margin: 0 }}>
-                  Expires {inviteExpiry.toLocaleDateString(undefined, { year: "numeric", month: "long", day: "numeric" })}
-                </p>
-              )}
-              <button
-                onClick={() => {
-                  setInviteStep("idle");
-                  setInviteToken(null);
-                  setInviteExpiry(null);
-                  setInviteDisplayName("");
-                  setInviteRole("Viewer");
-                }}
-                style={{ alignSelf: "flex-start", fontFamily: S.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", padding: "0.375rem 0.75rem", background: "none", border: `1px solid ${S.rule}`, color: S.inkLight, cursor: "pointer" }}
-              >
-                Invite Another →
-              </button>
-            </div>
-          )}
-
-        </div>
-      </div>
-
-    </div>
-  );
-}
-
-// ─── Rooms Tab ────────────────────────────────────────────────────────────────
-
-const FLOOR_TYPES = ["Hardwood", "Tile", "Carpet", "Laminate", "Vinyl", "Concrete", "Stone", "Other"];
-
-const EMPTY_FIXTURE_FORM: AddFixtureArgs = {
-  brand: "",
-  model: "",
-  serialNumber: "",
-  installedDate: "",
-  warrantyExpiry: "",
-  notes: "",
-};
-
-function RoomsTab({
-  propertyId,
-  rooms,
-  onRoomsChange,
-  photosByJob,
-  onRoomPhotoUpload,
-}: {
-  propertyId:         string;
-  rooms:              RoomRecord[];
-  onRoomsChange:      (rooms: RoomRecord[]) => void;
-  photosByJob:        Record<string, Photo[]>;
-  onRoomPhotoUpload:  (roomId: string, file: File) => Promise<void>;
-}) {
-  const [showAddRoom,    setShowAddRoom]    = useState(false);
-  const [savingRoom,     setSavingRoom]     = useState(false);
-  const [expandedRoom,   setExpandedRoom]   = useState<string | null>(null);
-  const [editingRoom,    setEditingRoom]    = useState<string | null>(null);
-  const [editForm,       setEditForm]       = useState<UpdateRoomArgs | null>(null);
-  const [addFixtureRoom, setAddFixtureRoom] = useState<string | null>(null);
-  const [fixtureForm,    setFixtureForm]    = useState<AddFixtureArgs>({ ...EMPTY_FIXTURE_FORM });
-  const [savingFixture,  setSavingFixture]  = useState(false);
-  const [uploadingPhoto, setUploadingPhoto] = useState<string | null>(null); // roomId being uploaded
-
-  const handleUpdateRoom = async (id: string) => {
-    if (!editForm || !editForm.name.trim()) return;
-    setSavingRoom(true);
-    try {
-      const updated = await roomService.updateRoom(id, editForm);
-      onRoomsChange(rooms.map((r) => r.id === id ? updated : r));
-      setEditingRoom(null);
-      setEditForm(null);
-    } catch (err: any) {
-      toast.error(err.message ?? "Failed to update room");
-    } finally {
-      setSavingRoom(false);
-    }
-  };
-
-  const handleDeleteRoom = async (id: string) => {
-    try {
-      await roomService.deleteRoom(id);
-      onRoomsChange(rooms.filter((r) => r.id !== id));
-      if (expandedRoom === id) setExpandedRoom(null);
-    } catch (err: any) {
-      toast.error(err.message ?? "Failed to delete room");
-    }
-  };
-
-  const handleAddFixture = async (roomId: string) => {
-    setSavingFixture(true);
-    try {
-      const updated = await roomService.addFixture(roomId, fixtureForm);
-      onRoomsChange(rooms.map((r) => r.id === roomId ? updated : r));
-      setAddFixtureRoom(null);
-      setFixtureForm({ ...EMPTY_FIXTURE_FORM });
-    } catch (err: any) {
-      toast.error(err.message ?? "Failed to add fixture");
-    } finally {
-      setSavingFixture(false);
-    }
-  };
-
-  const handleRemoveFixture = async (roomId: string, fixtureId: string) => {
-    try {
-      const updated = await roomService.removeFixture(roomId, fixtureId);
-      onRoomsChange(rooms.map((r) => r.id === roomId ? updated : r));
-    } catch (err: any) {
-      toast.error(err.message ?? "Failed to remove fixture");
-    }
-  };
-
-  const inputStyle: React.CSSProperties = {
-    width: "100%", padding: "0.5rem 0.625rem",
-    fontFamily: FONTS.sans, fontSize: "0.8rem",
-    border: `1px solid ${COLORS.rule}`, background: COLORS.white,
-    color: COLORS.plum, outline: "none",
-  };
-
-  const labelStyle: React.CSSProperties = {
-    display: "block", fontFamily: FONTS.mono, fontSize: "0.6rem",
-    letterSpacing: "0.08em", textTransform: "uppercase",
-    color: COLORS.plumMid, marginBottom: "0.25rem",
-  };
-
-  return (
-    <div>
-      {/* Add Room modal */}
-      <AddRoomModal
-        isOpen={showAddRoom}
-        onClose={() => setShowAddRoom(false)}
-        propertyId={propertyId}
-        onSuccess={(room) => {
-          onRoomsChange([...rooms, room]);
-          setExpandedRoom(room.id);
-        }}
-      />
-
-      {/* Header row */}
-      <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "1.25rem" }}>
-        <p style={{ fontFamily: FONTS.sans, fontSize: "0.85rem", color: COLORS.plumMid, fontWeight: 300 }}>
-          Track finishes, paint, and appliances room by room.
-        </p>
-        <Button size="sm" onClick={() => setShowAddRoom(true)}>+ Add Room</Button>
-      </div>
-
-      {/* Room list */}
-      {rooms.length === 0 && !showAddRoom && (
-        <div style={{ textAlign: "center", padding: "3rem 1rem", border: `1px dashed ${COLORS.rule}` }}>
-          <p style={{ fontFamily: FONTS.sans, color: COLORS.plumMid, fontSize: "0.875rem", fontWeight: 300 }}>
-            No rooms yet. Add your first room to start building your digital twin.
-          </p>
-        </div>
-      )}
-
-      <div style={{ display: "flex", flexDirection: "column", gap: "0.75rem" }}>
-        {rooms.map((room) => {
-          const isExpanded = expandedRoom === room.id;
-          const isEditing  = editingRoom  === room.id;
-
-          return (
-            <div key={room.id} style={{ border: `1px solid ${COLORS.rule}`, background: COLORS.white, boxShadow: SHADOWS.card }}>
-              {/* Room header */}
-              <div
-                style={{ display: "flex", alignItems: "center", justifyContent: "space-between", padding: "1rem 1.25rem", cursor: "pointer" }}
-                onClick={() => setExpandedRoom(isExpanded ? null : room.id)}
-              >
-                <div style={{ display: "flex", alignItems: "center", gap: "1rem" }}>
-                  <div>
-                    <div style={{ fontFamily: FONTS.serif, fontWeight: 700, fontSize: "1rem", color: COLORS.plum }}>{room.name}</div>
-                    <div style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.06em", color: COLORS.plumMid, marginTop: "0.2rem" }}>
-                      {[room.floorType, room.paintColor && `${room.paintColor}${room.paintCode ? ` (${room.paintCode})` : ""}`]
-                        .filter(Boolean).join(" · ") || "No finishes recorded"}
-                      {room.fixtures.length > 0 && ` · ${room.fixtures.length} fixture${room.fixtures.length === 1 ? "" : "s"}`}
-                    </div>
-                  </div>
-                </div>
-                <div style={{ display: "flex", alignItems: "center", gap: "0.5rem" }} onClick={(e) => e.stopPropagation()}>
-                  <button
-                    onClick={() => { setEditingRoom(room.id); setEditForm({ name: room.name, floorType: room.floorType, paintColor: room.paintColor, paintBrand: room.paintBrand, paintCode: room.paintCode, notes: room.notes }); setExpandedRoom(room.id); }}
-                    style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: COLORS.plumMid, background: "none", border: "none", cursor: "pointer", padding: "0.25rem 0.5rem" }}
-                  >Edit</button>
-                  <button
-                    onClick={() => { if (window.confirm(`Delete "${room.name}"?`)) handleDeleteRoom(room.id); }}
-                    style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: COLORS.sage, background: "none", border: "none", cursor: "pointer", padding: "0.25rem 0.5rem" }}
-                  >Delete</button>
-                  <span style={{ color: COLORS.plumMid, fontSize: "0.75rem" }}>{isExpanded ? "▲" : "▼"}</span>
-                </div>
-              </div>
-
-              {/* Expanded body */}
-              {isExpanded && (
-                <div style={{ borderTop: `1px solid ${COLORS.rule}`, padding: "1rem 1.25rem" }}>
-
-                  {/* Edit form */}
-                  {isEditing && editForm && (
-                    <div style={{ marginBottom: "1.25rem", padding: "1rem", background: COLORS.white, border: `1px solid ${COLORS.rule}` }}>
-                      <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "0.75rem", marginBottom: "0.75rem" }}>
-                        <div>
-                          <label style={labelStyle}>Room Name *</label>
-                          <input style={inputStyle} value={editForm.name}
-                            onChange={(e) => setEditForm((f) => f && ({ ...f, name: e.target.value }))} />
-                        </div>
-                        <div>
-                          <label style={labelStyle}>Floor Type</label>
-                          <select style={inputStyle} value={editForm.floorType}
-                            onChange={(e) => setEditForm((f) => f && ({ ...f, floorType: e.target.value }))}>
-                            <option value="">— Select —</option>
-                            {FLOOR_TYPES.map((t) => <option key={t} value={t}>{t}</option>)}
-                          </select>
-                        </div>
-                        <div>
-                          <label style={labelStyle}>Paint Color</label>
-                          <input style={inputStyle} value={editForm.paintColor}
-                            onChange={(e) => setEditForm((f) => f && ({ ...f, paintColor: e.target.value }))} />
-                        </div>
-                        <div>
-                          <label style={labelStyle}>Paint Brand</label>
-                          <input style={inputStyle} value={editForm.paintBrand}
-                            onChange={(e) => setEditForm((f) => f && ({ ...f, paintBrand: e.target.value }))} />
-                        </div>
-                        <div>
-                          <label style={labelStyle}>Paint Code</label>
-                          <input style={inputStyle} value={editForm.paintCode}
-                            onChange={(e) => setEditForm((f) => f && ({ ...f, paintCode: e.target.value }))} />
-                        </div>
-                      </div>
-                      <div style={{ marginBottom: "1rem" }}>
-                        <label style={labelStyle}>Notes</label>
-                        <textarea style={{ ...inputStyle, minHeight: "3.5rem", resize: "vertical" }} value={editForm.notes}
-                          onChange={(e) => setEditForm((f) => f && ({ ...f, notes: e.target.value }))} />
-                      </div>
-                      <div style={{ display: "flex", gap: "0.5rem" }}>
-                        <Button size="sm" onClick={() => handleUpdateRoom(room.id)} disabled={savingRoom}>
-                          {savingRoom ? "Saving…" : "Save Changes"}
-                        </Button>
-                        <Button size="sm" variant="outline" onClick={() => { setEditingRoom(null); setEditForm(null); }}>Cancel</Button>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Notes display */}
-                  {!isEditing && room.notes && (
-                    <p style={{ fontFamily: FONTS.sans, fontSize: "0.8rem", color: COLORS.plumMid, fontWeight: 300, marginBottom: "1rem", fontStyle: "italic" }}>
-                      {room.notes}
-                    </p>
-                  )}
-
-                  {/* Photo gallery */}
-                  {(() => {
-                    const roomPhotos = photosByJob[`ROOM_${room.id}`] ?? [];
-                    return (
-                      <div style={{ marginBottom: "1rem" }}>
-                        <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "0.5rem" }}>
-                          <span style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", color: COLORS.plumMid }}>
-                            Photos{roomPhotos.length > 0 ? ` (${roomPhotos.length})` : ""}
-                          </span>
-                          <label style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: COLORS.sage, cursor: "pointer" }}>
-                            {uploadingPhoto === room.id ? "Uploading…" : "+ Add Photo"}
-                            <input
-                              type="file"
-                              accept="image/*"
-                              style={{ display: "none" }}
-                              disabled={uploadingPhoto === room.id}
-                              onChange={async (e) => {
-                                const file = e.target.files?.[0];
-                                if (!file) return;
-                                setUploadingPhoto(room.id);
-                                try { await onRoomPhotoUpload(room.id, file); }
-                                finally { setUploadingPhoto(null); e.target.value = ""; }
-                              }}
-                            />
-                          </label>
-                        </div>
-                        {roomPhotos.length > 0 && (
-                          <div style={{ display: "flex", gap: "0.5rem", flexWrap: "wrap" }}>
-                            {roomPhotos.map((photo) => (
-                              <img
-                                key={photo.id}
-                                src={photo.url}
-                                alt={photo.description || "Room photo"}
-                                style={{ width: "5rem", height: "5rem", objectFit: "cover", border: `1px solid ${COLORS.rule}` }}
-                              />
-                            ))}
-                          </div>
-                        )}
-                        {roomPhotos.length === 0 && (
-                          <p style={{ fontFamily: FONTS.sans, fontSize: "0.75rem", color: COLORS.plumMid, fontWeight: 300, fontStyle: "italic" }}>
-                            No photos yet.
-                          </p>
-                        )}
-                      </div>
-                    );
-                  })()}
-
-                  {/* Fixtures */}
-                  <div style={{ marginBottom: "0.75rem" }}>
-                    <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "0.5rem" }}>
-                      <span style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase", color: COLORS.plumMid }}>
-                        Appliances & Fixtures
-                      </span>
-                      {addFixtureRoom !== room.id && (
-                        <button
-                          onClick={() => { setAddFixtureRoom(room.id); setFixtureForm({ ...EMPTY_FIXTURE_FORM }); }}
-                          style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.08em", textTransform: "uppercase", color: COLORS.sage, background: "none", border: "none", cursor: "pointer" }}
-                        >+ Add Fixture</button>
-                      )}
-                    </div>
-
-                    {/* Add fixture form */}
-                    {addFixtureRoom === room.id && (
-                      <div style={{ border: `1px solid ${COLORS.rule}`, padding: "1rem", marginBottom: "0.75rem", background: COLORS.white }}>
-                        <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "0.625rem", marginBottom: "0.625rem" }}>
-                          <div>
-                            <label style={labelStyle}>Brand</label>
-                            <input style={inputStyle} placeholder="e.g. KitchenAid" value={fixtureForm.brand}
-                              onChange={(e) => setFixtureForm((f) => ({ ...f, brand: e.target.value }))} />
-                          </div>
-                          <div>
-                            <label style={labelStyle}>Model</label>
-                            <input style={inputStyle} placeholder="e.g. KFIS29PBMS" value={fixtureForm.model}
-                              onChange={(e) => setFixtureForm((f) => ({ ...f, model: e.target.value }))} />
-                          </div>
-                          <div>
-                            <label style={labelStyle}>Serial Number</label>
-                            <input style={inputStyle} value={fixtureForm.serialNumber}
-                              onChange={(e) => setFixtureForm((f) => ({ ...f, serialNumber: e.target.value }))} />
-                          </div>
-                          <div>
-                            <label style={labelStyle}>Installed Date</label>
-                            <input style={inputStyle} type="date" value={fixtureForm.installedDate}
-                              onChange={(e) => setFixtureForm((f) => ({ ...f, installedDate: e.target.value }))} />
-                          </div>
-                          <div>
-                            <label style={labelStyle}>Warranty Expires</label>
-                            <input style={inputStyle} type="date" value={fixtureForm.warrantyExpiry}
-                              onChange={(e) => setFixtureForm((f) => ({ ...f, warrantyExpiry: e.target.value }))} />
-                          </div>
-                          <div>
-                            <label style={labelStyle}>Notes</label>
-                            <input style={inputStyle} placeholder="e.g. French door refrigerator" value={fixtureForm.notes}
-                              onChange={(e) => setFixtureForm((f) => ({ ...f, notes: e.target.value }))} />
-                          </div>
-                        </div>
-                        <div style={{ display: "flex", gap: "0.5rem" }}>
-                          <Button size="sm" onClick={() => handleAddFixture(room.id)} disabled={savingFixture}>
-                            {savingFixture ? "Saving…" : "Add Fixture"}
-                          </Button>
-                          <Button size="sm" variant="outline" onClick={() => setAddFixtureRoom(null)}>Cancel</Button>
-                        </div>
-                      </div>
-                    )}
-
-                    {/* Fixture list */}
-                    {room.fixtures.length === 0 && addFixtureRoom !== room.id && (
-                      <p style={{ fontFamily: FONTS.sans, fontSize: "0.75rem", color: COLORS.plumMid, fontWeight: 300, fontStyle: "italic" }}>
-                        No fixtures recorded.
-                      </p>
-                    )}
-                    {room.fixtures.map((f) => {
-                      const isExpired = f.warrantyExpiry && new Date(f.warrantyExpiry) < new Date();
-                      const expiringSoon = f.warrantyExpiry && !isExpired && new Date(f.warrantyExpiry) < new Date(Date.now() + 90 * 24 * 60 * 60 * 1000);
-                      return (
-                        <div key={f.id} style={{ display: "flex", alignItems: "flex-start", justifyContent: "space-between", padding: "0.625rem 0", borderBottom: `1px solid ${COLORS.rule}` }}>
-                          <div>
-                            <div style={{ fontFamily: FONTS.sans, fontWeight: 500, fontSize: "0.825rem", color: COLORS.plum }}>
-                              {f.brand} {f.model}
-                            </div>
-                            <div style={{ fontFamily: FONTS.mono, fontSize: "0.55rem", letterSpacing: "0.06em", color: COLORS.plumMid, marginTop: "0.2rem" }}>
-                              {f.serialNumber && `S/N ${f.serialNumber}`}
-                              {f.installedDate && ` · Installed ${f.installedDate}`}
-                              {f.warrantyExpiry && (
-                                <span style={{ color: isExpired ? COLORS.sage : expiringSoon ? "#C94C2E" : COLORS.plumMid }}>
-                                  {` · Warranty ${isExpired ? "expired" : "expires"} ${f.warrantyExpiry}`}
-                                </span>
-                              )}
-                            </div>
-                            {f.notes && (
-                              <div style={{ fontFamily: FONTS.sans, fontSize: "0.75rem", color: COLORS.plumMid, fontWeight: 300, marginTop: "0.2rem" }}>
-                                {f.notes}
-                              </div>
-                            )}
-                          </div>
-                          <button
-                            onClick={() => handleRemoveFixture(room.id, f.id)}
-                            style={{ fontFamily: FONTS.mono, fontSize: "0.55rem", letterSpacing: "0.06em", color: COLORS.plumMid, background: "none", border: "none", cursor: "pointer", flexShrink: 0, marginLeft: "1rem" }}
-                          >Remove</button>
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-// ─── Bills Tab ────────────────────────────────────────────────────────────────
-
-const BILL_TYPE_LABELS: Record<BillType, string> = {
-  Electric: "Electric",
-  Gas:      "Gas",
-  Water:    "Water",
-  Internet: "Internet",
-  Telecom:  "Telecom/Cable",
-  Other:    "Other",
-};
-
-function BillsTab({ propertyId }: { propertyId: string }) {
-  const [bills, setBills]                       = useState<BillRecord[]>([]);
-  const [loading, setLoading]                   = useState(true);
-  const [uploading, setUploading]               = useState(false);
-  const [extraction, setExtraction]             = useState<BillExtraction | null>(null);
-  const [confirmArgs, setConfirmArgs]           = useState<Partial<BillRecord> | null>(null);
-  const fileInputRef                            = React.useRef<HTMLInputElement>(null);
-
-  // Story 3 — efficiency alert state
-  const [efficiencyRec, setEfficiencyRec]       = useState<string | null>(null);
-  const [efficiencyWaste, setEfficiencyWaste]   = useState<number | null>(null);
-
-  // Story 4 — rebates state
-  const [rebates, setRebates]                   = useState<RebateResult[] | null>(null);
-  const [loadingRebates, setLoadingRebates]     = useState(false);
-
-  // Story 6 — telecom negotiation state
-  const [telecomResult, setTelecomResult]       = useState<TelecomNegotiationResult | null>(null);
-  const [telecomBillId, setTelecomBillId]       = useState<string | null>(null);
-  const [loadingTelecom, setLoadingTelecom]     = useState(false);
-
-  useEffect(() => {
-    billService.getBillsForProperty(propertyId)
-      .then(async (fetched) => {
-        setBills(fetched);
-
-        // Story 3 — run efficiency analysis for Electric bills when 3+ present
-        const elecTrend = await getUsageTrend(propertyId, "Electric", 12).catch(() => []);
-        const waterTrend = await getUsageTrend(propertyId, "Water", 12).catch(() => []);
-        const trend = elecTrend.length >= waterTrend.length ? elecTrend : waterTrend;
-        if (trend.length >= 3) {
-          const analysis = analyzeEfficiencyTrend(trend);
-          if (analysis.degradationDetected) {
-            setEfficiencyRec(analysis.recommendation ?? null);
-            setEfficiencyWaste(analysis.estimatedAnnualWaste ?? null);
-          }
-        }
-
-        // Story 4 — load rebates if Electric bills are present
-        const hasElectric = fetched.some((b) => b.billType === "Electric");
-        if (hasElectric) {
-          setLoadingRebates(true);
-          findRebates({ state: "FL", zipCode: "32801", utilityProvider: fetched.find((b) => b.billType === "Electric")?.provider ?? "Unknown", billType: "Electric" })
-            .then(setRebates)
-            .catch(() => {})
-            .finally(() => setLoadingRebates(false));
-        }
-      })
-      .catch(() => {})
-      .finally(() => setLoading(false));
-  }, [propertyId]);
-
-  async function handleFileChange(e: React.ChangeEvent<HTMLInputElement>) {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setUploading(true);
-    setExtraction(null);
-    try {
-      const base64Data = await fileToBase64(file);
-      const result = await extractBill(file.name, file.type, base64Data);
-      setExtraction(result);
-      setConfirmArgs({
-        billType:    result.billType ?? "Other",
-        provider:    result.provider ?? "",
-        periodStart: result.periodStart ?? "",
-        periodEnd:   result.periodEnd ?? "",
-        amountCents: result.amountCents ?? 0,
-        usageAmount: result.usageAmount,
-        usageUnit:   result.usageUnit,
-      });
-    } catch {
-      toast.error("Could not extract bill data. Please fill in the details manually.");
-      setConfirmArgs({ billType: "Other", provider: "", periodStart: "", periodEnd: "", amountCents: 0 });
-    } finally {
-      setUploading(false);
-      if (fileInputRef.current) fileInputRef.current.value = "";
-    }
-  }
-
-  async function handleConfirmSave() {
-    if (!confirmArgs) return;
-    try {
-      const saved = await billService.addBill({
-        propertyId,
-        billType:    (confirmArgs.billType as BillType) ?? "Other",
-        provider:    confirmArgs.provider ?? "",
-        periodStart: confirmArgs.periodStart ?? "",
-        periodEnd:   confirmArgs.periodEnd ?? "",
-        amountCents: confirmArgs.amountCents ?? 0,
-        usageAmount: confirmArgs.usageAmount,
-        usageUnit:   confirmArgs.usageUnit,
-      });
-      setBills((prev) => [saved, ...prev]);
-      setConfirmArgs(null);
-      setExtraction(null);
-      toast.success("Bill saved.");
-      if (saved.anomalyFlag) {
-        toast(`Anomaly detected: ${saved.anomalyReason}`, { icon: "⚠️", duration: 6000 });
-      }
-    } catch (err) {
-      if (err instanceof TierLimitReachedError) {
-        toast.error(err.message, { duration: 8000, icon: "🔒" });
-      } else {
-        toast.error("Failed to save bill.");
-      }
-    }
-  }
-
-  async function handleDelete(id: string) {
-    try {
-      await billService.deleteBill(id);
-      setBills((prev) => prev.filter((b) => b.id !== id));
-      toast.success("Bill removed.");
-    } catch {
-      toast.error("Failed to remove bill.");
-    }
-  }
-
-  // Story 6 — Telecom Negotiation
-  async function handleNegotiateTelecom(bill: BillRecord) {
-    if (!bill.amountCents) return;
-    setLoadingTelecom(true);
-    setTelecomBillId(bill.id);
-    try {
-      const result = await negotiateTelecom({
-        provider:    bill.provider,
-        amountCents: bill.amountCents,
-        mbps:        bill.usageAmount ?? 100,
-        zipCode:     "32801",
-      });
-      setTelecomResult(result);
-    } catch {
-      toast.error("Could not generate negotiation script. Try again later.");
-    } finally {
-      setLoadingTelecom(false);
-    }
-  }
-
-  const ink      = COLORS.plum;
-  const rule     = COLORS.rule;
-  const inkLight = COLORS.plumMid;
-
-  return (
-    <div style={{ padding: "2rem 0" }}>
-      <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "1.5rem" }}>
-        <div>
-          <h3 style={{ fontFamily: FONTS.serif, fontSize: "1.25rem", fontWeight: 700, color: ink, margin: 0 }}>
-            Utility Bills
-          </h3>
-          <p style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", color: inkLight, margin: "0.25rem 0 0" }}>
-            Upload bills to track usage, detect anomalies, and surface savings opportunities.
-          </p>
-        </div>
-        <Button onClick={() => fileInputRef.current?.click()} disabled={uploading}>
-          <Upload size={14} style={{ marginRight: "0.4rem" }} />
-          {uploading ? "Extracting…" : "Upload Bill"}
-        </Button>
-        <input
-          ref={fileInputRef}
-          type="file"
-          accept="image/jpeg,image/png,image/webp,image/gif,application/pdf"
-          style={{ display: "none" }}
-          onChange={handleFileChange}
-        />
-      </div>
-
-      {confirmArgs && (
-        <div style={{ border: `1px solid ${rule}`, background: COLORS.white, padding: "1.5rem", marginBottom: "1.5rem" }}>
-          <h4 style={{ fontFamily: FONTS.serif, fontSize: "1rem", fontWeight: 700, color: ink, margin: "0 0 0.25rem" }}>
-            Confirm Bill Details
-          </h4>
-          {extraction && (
-            <p style={{ fontFamily: FONTS.sans, fontSize: "0.8rem", color: inkLight, margin: "0 0 1rem" }}>
-              {extraction.description} (confidence: {extraction.confidence})
-            </p>
-          )}
-          <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "1rem" }}>
-            <label style={{ display: "flex", flexDirection: "column", gap: "0.25rem" }}>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.65rem", textTransform: "uppercase" as const, letterSpacing: "0.08em", color: inkLight }}>Bill Type</span>
-              <select
-                value={confirmArgs.billType ?? "Other"}
-                onChange={(e) => setConfirmArgs((p) => ({ ...p!, billType: e.target.value as BillType }))}
-                style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", padding: "0.4rem 0.5rem", border: `1px solid ${rule}`, color: ink, background: COLORS.white }}
-              >
-                {(Object.keys(BILL_TYPE_LABELS) as BillType[]).map((t) => (
-                  <option key={t} value={t}>{BILL_TYPE_LABELS[t]}</option>
-                ))}
-              </select>
-            </label>
-            <label style={{ display: "flex", flexDirection: "column", gap: "0.25rem" }}>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.65rem", textTransform: "uppercase" as const, letterSpacing: "0.08em", color: inkLight }}>Provider</span>
-              <input
-                type="text"
-                value={confirmArgs.provider ?? ""}
-                onChange={(e) => setConfirmArgs((p) => ({ ...p!, provider: e.target.value }))}
-                placeholder="e.g. FPL, TECO"
-                style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", padding: "0.4rem 0.5rem", border: `1px solid ${rule}`, color: ink, background: COLORS.white }}
-              />
-            </label>
-            <label style={{ display: "flex", flexDirection: "column", gap: "0.25rem" }}>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.65rem", textTransform: "uppercase" as const, letterSpacing: "0.08em", color: inkLight }}>Period Start</span>
-              <input
-                type="date"
-                value={confirmArgs.periodStart ?? ""}
-                onChange={(e) => setConfirmArgs((p) => ({ ...p!, periodStart: e.target.value }))}
-                style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", padding: "0.4rem 0.5rem", border: `1px solid ${rule}`, color: ink, background: COLORS.white }}
-              />
-            </label>
-            <label style={{ display: "flex", flexDirection: "column", gap: "0.25rem" }}>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.65rem", textTransform: "uppercase" as const, letterSpacing: "0.08em", color: inkLight }}>Period End</span>
-              <input
-                type="date"
-                value={confirmArgs.periodEnd ?? ""}
-                onChange={(e) => setConfirmArgs((p) => ({ ...p!, periodEnd: e.target.value }))}
-                style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", padding: "0.4rem 0.5rem", border: `1px solid ${rule}`, color: ink, background: COLORS.white }}
-              />
-            </label>
-            <label style={{ display: "flex", flexDirection: "column", gap: "0.25rem" }}>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.65rem", textTransform: "uppercase" as const, letterSpacing: "0.08em", color: inkLight }}>Amount ($)</span>
-              <input
-                type="number"
-                min={0}
-                step={0.01}
-                value={confirmArgs.amountCents != null ? (confirmArgs.amountCents / 100).toFixed(2) : ""}
-                onChange={(e) => setConfirmArgs((p) => ({ ...p!, amountCents: Math.round(parseFloat(e.target.value) * 100) || 0 }))}
-                placeholder="0.00"
-                style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", padding: "0.4rem 0.5rem", border: `1px solid ${rule}`, color: ink, background: COLORS.white }}
-              />
-            </label>
-            <label style={{ display: "flex", flexDirection: "column", gap: "0.25rem" }}>
-              <span style={{ fontFamily: FONTS.mono, fontSize: "0.65rem", textTransform: "uppercase" as const, letterSpacing: "0.08em", color: inkLight }}>Usage (optional)</span>
-              <div style={{ display: "flex", gap: "0.5rem" }}>
-                <input
-                  type="number"
-                  min={0}
-                  value={confirmArgs.usageAmount ?? ""}
-                  onChange={(e) => setConfirmArgs((p) => ({ ...p!, usageAmount: e.target.value ? parseFloat(e.target.value) : undefined }))}
-                  placeholder="842"
-                  style={{ flex: 1, fontFamily: FONTS.sans, fontSize: "0.875rem", padding: "0.4rem 0.5rem", border: `1px solid ${rule}`, color: ink, background: COLORS.white }}
-                />
-                <select
-                  value={confirmArgs.usageUnit ?? "kWh"}
-                  onChange={(e) => setConfirmArgs((p) => ({ ...p!, usageUnit: e.target.value }))}
-                  style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", padding: "0.4rem 0.5rem", border: `1px solid ${rule}`, color: ink, background: COLORS.white }}
-                >
-                  <option value="kWh">kWh</option>
-                  <option value="gallons">gallons</option>
-                  <option value="therms">therms</option>
-                  <option value="Mbps">Mbps</option>
-                </select>
-              </div>
-            </label>
-          </div>
-          <div style={{ display: "flex", gap: "0.75rem", marginTop: "1.25rem" }}>
-            <Button onClick={handleConfirmSave}>Save Bill</Button>
-            <Button
-              onClick={() => { setConfirmArgs(null); setExtraction(null); }}
-              style={{ background: "transparent", border: `1px solid ${rule}`, color: inkLight }}
-            >
-              Cancel
-            </Button>
-          </div>
-        </div>
-      )}
-
-      {loading ? (
-        <div style={{ display: "flex", justifyContent: "center", padding: "3rem" }}>
-          <div className="spinner-lg" />
-        </div>
-      ) : bills.length === 0 && !confirmArgs ? (
-        <div style={{ textAlign: "center", padding: "3rem", color: inkLight, fontFamily: FONTS.sans, fontSize: "0.9rem" }}>
-          No bills uploaded yet. Upload a utility bill to start tracking.
-        </div>
-      ) : (
-        <table style={{ width: "100%", borderCollapse: "collapse" }}>
-          <thead>
-            <tr style={{ borderBottom: `1px solid ${rule}` }}>
-              {["Type", "Provider", "Period", "Amount", "Usage", ""].map((h) => (
-                <th key={h} style={{ fontFamily: FONTS.mono, fontSize: "0.65rem", textTransform: "uppercase" as const, letterSpacing: "0.08em", color: inkLight, textAlign: "left", padding: "0.5rem 0.75rem", fontWeight: 500 }}>
-                  {h}
-                </th>
-              ))}
-            </tr>
-          </thead>
-          <tbody>
-            {[...bills].sort((a, b) => b.uploadedAt - a.uploadedAt).map((bill) => (
-              <tr key={bill.id} style={{ borderBottom: `1px solid ${rule}` }}>
-                <td style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", color: ink, padding: "0.75rem" }}>
-                  {BILL_TYPE_LABELS[bill.billType]}
-                </td>
-                <td style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", color: ink, padding: "0.75rem" }}>
-                  {bill.provider}
-                </td>
-                <td style={{ fontFamily: FONTS.mono, fontSize: "0.8rem", color: inkLight, padding: "0.75rem", whiteSpace: "nowrap" as const }}>
-                  {bill.periodStart} → {bill.periodEnd}
-                </td>
-                <td style={{ fontFamily: FONTS.sans, fontSize: "0.875rem", color: ink, padding: "0.75rem", fontWeight: 600 }}>
-                  ${(bill.amountCents / 100).toFixed(2)}
-                </td>
-                <td style={{ fontFamily: FONTS.mono, fontSize: "0.8rem", color: inkLight, padding: "0.75rem" }}>
-                  {bill.usageAmount != null && bill.usageUnit
-                    ? `${bill.usageAmount.toLocaleString()} ${bill.usageUnit}`
-                    : "—"}
-                </td>
-                <td style={{ padding: "0.75rem", whiteSpace: "nowrap" as const }}>
-                  {bill.anomalyFlag && (
-                    <span
-                      title={bill.anomalyReason}
-                      style={{ display: "inline-flex", alignItems: "center", gap: "0.25rem", fontFamily: FONTS.mono, fontSize: "0.7rem", color: "#C94C2E", marginRight: "0.75rem" }}
-                    >
-                      <Zap size={12} /> Anomaly
-                    </span>
-                  )}
-                  {(bill.billType === "Internet" || bill.billType === "Telecom") && (
-                    <button
-                      onClick={() => handleNegotiateTelecom(bill)}
-                      disabled={loadingTelecom && telecomBillId === bill.id}
-                      title="Negotiate your bill"
-                      style={{ display: "inline-flex", alignItems: "center", gap: "0.25rem", background: "none", border: `1px solid ${COLORS.sage}`, borderRadius: "4px", cursor: "pointer", color: COLORS.sage, padding: "0.2rem 0.5rem", fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.06em", marginRight: "0.5rem" }}
-                    >
-                      <PhoneCall size={10} />
-                      {loadingTelecom && telecomBillId === bill.id ? "…" : "Negotiate"}
-                    </button>
-                  )}
-                  <button
-                    onClick={() => handleDelete(bill.id)}
-                    style={{ background: "none", border: "none", cursor: "pointer", color: inkLight, padding: "0.25rem" }}
-                    title="Remove bill"
-                  >
-                    <Trash2 size={14} />
-                  </button>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      )}
-
-      {/* Story 3 — Efficiency Degradation Alert */}
-      {efficiencyRec && (
-        <div style={{ marginTop: "1.5rem", padding: "1.25rem 1.5rem", background: COLORS.butter, border: `1px solid ${COLORS.rule}`, borderRadius: RADIUS.card, display: "flex", gap: "1rem", alignItems: "flex-start" }}>
-          <TrendingUp size={18} color={COLORS.plum} style={{ flexShrink: 0, marginTop: "0.15rem" }} />
-          <div>
-            <p style={{ fontFamily: FONTS.serif, fontWeight: 700, fontSize: "1rem", color: COLORS.plum, margin: "0 0 0.375rem" }}>
-              Usage trend detected
-            </p>
-            <p style={{ fontFamily: FONTS.sans, fontSize: "0.85rem", fontWeight: 300, color: COLORS.plumMid, margin: 0, lineHeight: 1.6 }}>
-              {efficiencyRec}
-              {efficiencyWaste != null && (
-                <> Estimated annual waste: <strong>{efficiencyWaste.toLocaleString(undefined, { maximumFractionDigits: 0 })} units</strong>.</>
-              )}
-            </p>
-          </div>
-        </div>
-      )}
-
-      {/* Story 4 — Available Rebates */}
-      {(loadingRebates || (rebates && rebates.length > 0)) && (
-        <div style={{ marginTop: "1.5rem" }}>
-          <p style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.1em", textTransform: "uppercase" as const, color: COLORS.plumMid, marginBottom: "0.75rem" }}>
-            Available Rebates
-          </p>
-          {loadingRebates ? (
-            <p style={{ fontFamily: FONTS.sans, fontSize: "0.85rem", color: COLORS.plumMid }}>Loading rebate programs…</p>
-          ) : (
-            <div style={{ display: "flex", flexDirection: "column", gap: "0.625rem" }}>
-              {rebates!.map((r) => (
-                <div key={r.name} style={{ padding: "1rem 1.25rem", background: COLORS.sageLight, border: `1px solid ${COLORS.rule}`, borderRadius: RADIUS.sm, display: "flex", justifyContent: "space-between", alignItems: "flex-start", gap: "1rem" }}>
-                  <div>
-                    <p style={{ fontFamily: FONTS.sans, fontWeight: 600, fontSize: "0.875rem", color: COLORS.plum, margin: "0 0 0.25rem" }}>{r.name}</p>
-                    <p style={{ fontFamily: FONTS.sans, fontWeight: 300, fontSize: "0.8rem", color: COLORS.plumMid, margin: 0, lineHeight: 1.55 }}>{r.description}</p>
-                    <p style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.06em", color: COLORS.plumMid, margin: "0.375rem 0 0" }}>{r.provider}</p>
-                  </div>
-                  <div style={{ flexShrink: 0, textAlign: "right" }}>
-                    <p style={{ fontFamily: FONTS.serif, fontWeight: 700, fontSize: "1rem", color: COLORS.sage, margin: "0 0 0.25rem" }}>{r.estimatedAmount}</p>
-                    {r.url && (
-                      <a href={r.url} target="_blank" rel="noopener noreferrer" style={{ display: "inline-flex", alignItems: "center", gap: "0.25rem", fontFamily: FONTS.mono, fontSize: "0.6rem", color: COLORS.sage, textDecoration: "none" }}>
-                        Apply <ExternalLink size={10} />
-                      </a>
-                    )}
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* Story 6 — Telecom Negotiation Result */}
-      {telecomResult && (
-        <div style={{ marginTop: "1.5rem", padding: "1.25rem 1.5rem", background: COLORS.white, border: `1px solid ${COLORS.sage}`, borderRadius: RADIUS.card }}>
-          <div style={{ display: "flex", alignItems: "center", gap: "0.625rem", marginBottom: "0.75rem" }}>
-            <PhoneCall size={16} color={COLORS.sage} />
-            <p style={{ fontFamily: FONTS.serif, fontWeight: 700, fontSize: "1rem", color: COLORS.plum, margin: 0 }}>
-              Negotiation Script
-            </p>
-            <span style={{ fontFamily: FONTS.mono, fontSize: "0.6rem", letterSpacing: "0.06em", padding: "2px 8px", borderRadius: "100px", background: telecomResult.verdict === "overpaying" ? "#FEE2E2" : COLORS.sageLight, color: telecomResult.verdict === "overpaying" ? "#C94C2E" : COLORS.sage }}>
-              {telecomResult.verdict === "overpaying" ? "Overpaying" : telecomResult.verdict === "fair" ? "Fair rate" : "Good deal"}
-            </span>
-          </div>
-          {telecomResult.savingsOpportunityCents > 0 && (
-            <p style={{ fontFamily: FONTS.sans, fontSize: "0.85rem", color: COLORS.plumMid, margin: "0 0 0.75rem" }}>
-              You may be paying <strong>${(telecomResult.savingsOpportunityCents / 100).toFixed(0)}/mo</strong> above the median rate of <strong>${(telecomResult.medianCents / 100).toFixed(0)}/mo</strong> for your area.
-            </p>
-          )}
-          <div style={{ padding: "1rem", background: COLORS.sageLight, borderRadius: RADIUS.sm, fontFamily: FONTS.sans, fontSize: "0.875rem", color: COLORS.plum, lineHeight: 1.7, whiteSpace: "pre-wrap" as const }}>
-            {telecomResult.negotiationScript}
-          </div>
-          <button onClick={() => setTelecomResult(null)} style={{ marginTop: "0.75rem", background: "none", border: "none", cursor: "pointer", fontFamily: FONTS.mono, fontSize: "0.6rem", color: COLORS.plumMid, letterSpacing: "0.06em" }}>
-            Dismiss
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function fileToBase64(file: File): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = () => {
-      const base64 = (reader.result as string).split(",")[1];
-      resolve(base64);
-    };
-    reader.onerror = reject;
-    reader.readAsDataURL(file);
-  });
-}
