+++ frontend/src/services/referralService.ts
@@ -29,7 +29,7 @@ export const referralService = {
   /** Fetch pending referral fees (admin). Returns empty array when canister absent. */
   async getPendingFees(): Promise<ReferralFeeRecord[]> {
     if (!JOB_CANISTER_ID) return [];
-    // TODO: call job canister getReferralFees() once implemented on-chain
+    // TODO(#82): call job canister getReferralFees() once implemented on-chain
     return [];
   },
 };
